### 2022-04-27 #4356 @jgostick
Hi @danielskatz, after a few bad review experiences earlier this year I swore I would avoid doing these "molecular dynamics" related packages. I just don't know enough about them to properly mediate disputes.  This package looks in great shape, but is also by its own admission "lightweight", meant for "method development", and under "substantial development".  All these things make me think this may not be a straightforward review.
