### 2022-04-27 0021fca
Corrected log_priori factor in table flip canonical swap case to 0.
