### 2022-04-28 #4356 @jarvist
Hi @TomTranter - could you review this Cluster Expansion package based in the pymatgen project? It seems to have a pretty good overlap with your public repositories.
