### 2022-04-28 #4356 @MarDiehl
> Similarly, @MarDiehl, would you be willing to review? There's perhaps less overlap in terms of language, but I would suppose you are quite aware of the maths / theory.

I'm familiar with Python but the methods differ a lot from what I'm doing. I think I can still provide helpful feedback, but considering also that it is my first review for JOSS I would appreciate if someone else takes the lead while I get the chance to get to know the process.
