### 2022-04-28 4bb4221
small doc edits (notebooks still not build correctly)
