+++ .github/workflows/test.yml
@@ -44,7 +44,7 @@ jobs:
           pytest tests --cov=smol --cov-report=xml
 
       - name: codacy-coverage-reporter
-        if: matrix.python-version == 3.9
+        if: ${{ matrix.python_version == 3.9 }}
         uses: codacy/codacy-coverage-reporter-action@v1
         with:
             project-token: ${{ secrets.CODACY_PROJECT_TOKEN }}
