### 2022-04-29 #4356 @TomTranter
Hi @jarvist I am not familiar with pymatgen at all and don't really do atomistic modelling. I know python very well though and can follow the examples. Can you try others on the list first then come back to me if you're struggling. Thanks
