### 2022-04-30 bb4bd64
prerun + fix example notebooks for docs
