### 2022-05-02 be2d886
[pre-commit.ci] pre-commit autoupdate
updates:
- https://gitlab.com/pycqa/flake8 → https://github.com/PyCQA/flake8
- [github.com/PyCQA/flake8: 3.9.2 → 4.0.1](https://github.com/PyCQA/flake8/compare/3.9.2...4.0.1)
