+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy>=1.20.1
-monty==2022.3.12
+monty==2022.4.26
 pymatgen==2022.4.26
