### 2022-05-13 4f6d3c6
add "import Lattice" to getting_started.rst
