+++ docs/src/getting_started.rst
@@ -28,7 +28,7 @@ Start by creating a disordered primitive structure.
 
 .. nbplot::
 
-    >>> from pymatgen.core.structure import Structure
+    >>> from pymatgen.core.structure import Structure, Lattice
     >>> species = {"Au": 0.5, "Cu": 0.5}
     >>> prim = Structure.from_spacegroup(
             "Fm-3m", Lattice.cubic(3.6), [species], [[0, 0, 0]])
