+++ docs/src/citing.rst
@@ -4,9 +4,10 @@ Citing
 
 If you use **smol** for your work please cite the following publication,
 
-    Barroso-Luque, L., Yang, J.H., Xie, F., Chen T., Zhong, P. & Ceder, G.
-    **smol**: A Python package for cluster expansions and beyond.
-    (submitted)
+    Barroso-Luque, L., Yang, J.H., Xie, F., Chen T., Kam, R.L., Jadidi, Z., Zhong, P.
+    & Ceder, G.
+    `smol: A Python package for cluster expansions and beyond. (submitted)
+    <https://joss.theoj.org/papers/e96a568ca53ee9d14548d7b8bed69b25>`_
 
 Since many core objects are used extensively, please also consider citing the
 **pymatgen** publication,
@@ -37,7 +38,7 @@ model, please cite the following,
 
     Yang, J.H., Chen, T., Barroso-Luque, L., Jadidi, Z. & Ceder, G.
     Approaches for handling high-dimensional cluster expansions of ionic
-    systems. (in preparation)
+    systems. (submitted)
 
     Barroso-Luque, L., Zhong, P., Yang, J. H., Chen, T. & Ceder, G.
     Cluster Expansions of Multicomponent Ionic Materials:Formalism & Methods.
@@ -50,5 +51,6 @@ Similarly, if you use :math:`\ell_0\ell_2` mixed integer quadratic programming
 for hierarchical constrained expansion fits, please cite this publication,
 
     Zhong, P., Chen, T., Barroso-Luque, L., Xie, F. & Ceder, G.
-    An :math:`\ell_0\ell_2`norm regularized regression model for construction
-    of robust cluster expansion in multicomponent systems. (in preparation)
+    An :math:`\ell_0\ell_2` norm regularized regression model for construction
+    of robust cluster expansion in multicomponent systems.
+    `[arxiv] <https://arxiv.org/abs/2204.13789>`_
