+++ docs/src/user_guide/index.rst
@@ -26,18 +26,20 @@ of the package is given in the :doc:`Developing </developer_guide/index>` sectio
 .. image:: ../../_static/use_workflow.png
    :width: 800px
 
+----------------------------------------------------------------------------------------
+
+============
+Main classes
+============
 
 Below is a general description of the core classes in each submodule, to help understand
 the design, usage and capabilities of **smol**. You can also refer to the :ref:`api ref`
 for full documentation of all classes and functions in the package.
 
-----------------------------------------------------------------------------------------
-
 .. _smol.cofe ug:
 
-=========
 smol.cofe
-=========
+---------
 
 This module includes the necessary classes to define, train, and test cluster
 expansions. A cluster expansion is essentially a way to fit a function of
@@ -58,7 +60,7 @@ The core classes are:
 .. _cluster subspace ug:
 
 Cluster subspace
-----------------
+^^^^^^^^^^^^^^^^
 :class:`ClusterSubspace` contains the finite set of orbits and orbit basis
 functions to be included in the cluster expansion.
 In general, a cluster expansion is created by first generating a
@@ -86,7 +88,7 @@ Full documentation of the class is available here, :ref:`cluster space`.
 .. _structure wrangler ug:
 
 Structure wrangler
-------------------
+^^^^^^^^^^^^^^^^^^
 :class:`StructureWrangler` handles input data structures and properties
 to fit to the cluster expansion.
 Once a set of structures and their relevant properties (for example, their
@@ -103,7 +105,7 @@ Full documentation of the class is available here: :ref:`structure wrangler`.
 .. _cluster expansion ug:
 
 Cluster expansion
------------------
+^^^^^^^^^^^^^^^^^
 :class:`ClusterExpansion` contains the fitted coefficents of the cluster
 expansion for predicting CE properties of new structures.
 Based on the feature matrix from the :class:`StructureWrangler`, one can fit
@@ -128,9 +130,8 @@ Full documentation of the class is available here: :ref:`cluster expansion`.
 
 .. _smol.moca ug:
 
-=========
 smol.moca
-=========
+---------
 
 This module includes classes and functions to run Markov Chain Monte Carlo
 sampling of statistical mechanical ensembles represented by a cluster expansion
@@ -163,7 +164,7 @@ The core classes are:
 .. _processors ug:
 
 Processors
-----------
+^^^^^^^^^^
 A :class:`Processor` is used to optimally compute correlation vectors, energy,
 and differences in these from variations in site occupancies. Processors
 compute values only for a specific supercell specified by a given supercell
@@ -182,7 +183,7 @@ Full documentation of the class and its subclasses available here: :ref:`process
 .. _ensembles ug:
 
 Ensembles
----------
+^^^^^^^^^
 :class:`Ensemble` classes represent the specific statistical mechanics ensemble
 by defining the relevant thermodynamic boundary conditions in order to compute
 the appropriate ensemble probability ratios. For example,
@@ -199,7 +200,7 @@ Full documentation of the class and its subclasses are available here: :ref:`ens
 .. _sampler ug:
 
 Sampler
--------
+^^^^^^^
 A :class:`Sampler` takes care of running MCMC sampling for a given ensemble.
 The easiest way to create a sampler (which suffices for most use cases) is to
 use the :meth:`from_ensemble` class method, which is sufficient for most cases using
@@ -214,7 +215,7 @@ Full documentation of the class is available here: :ref:`sampler`.
 .. _samplecontainer ug:
 
 SampleContainer
----------------
+^^^^^^^^^^^^^^^
 A :class:`SampleContainer` stores data from Monte Carlo sampling simulations,
 especially the occupancies and feature vectors. For lenghty MC simulations a
 :class:`SampleContainer` allows streaming directly to an
