+++ requirements.txt
@@ -1,3 +1,7 @@
 numpy>=1.20.1
 monty==2022.4.26
-pymatgen==2022.4.26
+pymatgen>=2022.4.26
+polytope
+cvxpy
+cvxopt
+h5py
