+++ requirements-optional.txt
@@ -1,4 +1,2 @@
-scikit-learn==1.1.0
-h5py==3.6.0
-cvxopt>=1.2.7
+scikit-learn>=1.1.0
 gurobipy
