### 2022-05-22 f55ce71
pass down single rng in moca
