### 2022-05-22 bee058f
rng use in composition select and fix sampler from_ensemble
