### 2022-05-23 #4356 @lbluque
@jarvist, here's a list sorted by (imo) relevance:
 - Quantumstud, ml-evs (have cluster expansion publications)
 - bocklund (pymatgen, calphad---which is quite relevant to CE)
 - wcwitt, srmnitc (atomic cluster expansion, not the same but related)
 - dandavies99, mturiansky, mbkumar  (pymatgen)
 - rosecers (stat physics, regression)
 - georgiastuart (MCMC)
