### 2022-05-23 842e3e9
add drx example to docs
