### 2022-05-23 85129bc
Supressed table flip warnings when mckernel initialize trace with an all-0 occupancy.
