### 2022-05-23 2879059
add drx example to docs
