+++ src/cemc_utils.pyx
@@ -58,38 +58,38 @@ cpdef corr_from_occupancy(const long[::1] occu,
     return out
 
 
-cpdef factors_from_occupancy(const long[::1] occu,
-                             const int num_factors,
+cpdef interactions_from_occupancy(const long[::1] occu,
+                             const int num_interactions,
                              const double offset,
                              list orbit_list):
-    """Computes the orbit factor vector for a given encoded occupancy string.
+    """Computes the cluster interaction vector for a given encoded occupancy string.
 
     Args:
         occu (ndarray):
             encoded occupancy vector
-        num_factors (int):
-            total number of bit orderings in expansion.
+        num_interactions (int):
+            total number of cluster interactions (orbits in cluster subspace).
         offset (float):
             eci value for the constant term.
         orbit_list:
             Information of all orbits.
-            (flat tensor index array, flat factor tensor,
+            (flat tensor index array, flat cluster interaction tensor,
              site indices of clusters)
 
     Returns: array
-        orbit factor vector for given occupancy
+        cluster interaction vector for given occupancy
     """
     cdef int n, i, j, I, J, index
     cdef double p
     cdef const long[:, ::1] indices
     cdef const long[::1] tensor_indices
-    cdef const double[::1] factor_tensors
-    out = np.zeros(num_factors)
+    cdef const double[::1] interaction_tensors
+    out = np.zeros(num_interactions)
     cdef double[:] o_view = out
     o_view[0] = offset  # empty cluster
 
     n = 1
-    for tensor_indices, factor_tensors, indices in orbit_list:
+    for tensor_indices, interaction_tensors, indices in orbit_list:
         I = indices.shape[0] # cluster index
         J = indices.shape[1] # index within cluster
         p = 0
@@ -97,7 +97,7 @@ cpdef factors_from_occupancy(const long[::1] occu,
             index = 0
             for j in range(J):
                 index += tensor_indices[j] * occu[indices[i, j]]
-            p += factor_tensors[index]
+            p += interaction_tensors[index]
         o_view[n] = p / I
         n += 1
 
@@ -211,11 +211,11 @@ cpdef indicator_delta_corr_single_flip(const long[::1] occu_f,
     return out
 
 
-cpdef delta_factors_single_flip(const long[::1] occu_f,
+cpdef delta_interactions_single_flip(const long[::1] occu_f,
                                 const long[::1] occu_i,
-                                const int num_factors,
+                                const int num_interactions,
                                 list site_orbit_list):
-    """Computes the orbit factor vector difference between two occupancy
+    """Computes the cluster interaction vector difference between two occupancy
     strings.
 
     Args:
@@ -223,26 +223,26 @@ cpdef delta_factors_single_flip(const long[::1] occu_f,
             encoded occupancy vector with flip
         occu_i (ndarray):
             encoded occupancy vector without flip
-        num_factors (int):
-            total number of bit orderings in expansion.
+        num_interactions (int):
+            total number of cluster interactions (orbits in cluster subspace).
         site_orbit_list:
             Information of all orbits that include the flip site.
             List of tuples each with
-            (cluster ratio, flat tensor index array, flat factor tensor,
+            (cluster ratio, flat tensor index array, flat cluster interaction tensor,
              site indices of clusters)
 
     Returns:
-        ndarray: orbit factor vector difference
+        ndarray: cluster interaction vector difference
     """
     cdef int i, j, n, I, J, ind_i, ind_f
     cdef double p, ratio
     cdef const long[:, ::1] indices
[…]
