+++ smol/moca/ensemble/ensemble.py
@@ -1,0 +1,368 @@
+"""Implementation of Thermodynamic Ensemble for MC sampling."""
+
+__author__ = "Luis Barroso-Luque"
+
+from collections import Counter
+
+import numpy as np
+from monty.json import MSONable
+
+from smol.cofe.space.domain import get_species
+from smol.moca.processor import (
+    ClusterExpansionProcessor,
+    CompositeProcessor,
+    EwaldProcessor,
+)
+from smol.moca.processor.base import Processor
+from smol.moca.sublattice import Sublattice
+
+
+class ChemicalPotentials:
+    """Chemical potential descriptor."""
+
+    name: str = "chemical_potentials"
+    natural_parameter: float = -1.0
+
+    def __init__(self):
+        """Initialize boundary descriptor."""
+        self._table = None
+        self._value = None
+
+    def __get__(self, obj, objtype=None):
+        """Return the table."""
+        return self._table
+
+    def __set__(self, obj, value):
+        """Set the table given the owner and value."""
+        # if first instantiation concatenate the natural parameter
+        if self._table is None:
+            obj.natural_parameters = np.append(
+                obj.natural_parameters, self.natural_parameter
+            )
+        # update the ensemble dictionary and _boundaries list
+        if hasattr(obj, "thermo_boundaries"):
+            obj.thermo_boundaries.update({self.name: value})
+        else:
+            setattr(obj, "thermo_boundaries", {self.name: value})
+
+        self._table = self._build_table(obj, value)
+        self._value = value
+
+    def __delete__(self, obj):
+        """Delete the boundary condition."""
+        self._table = None
+        self._value = None
+
+        if hasattr(obj, "thermo_boundaries"):
+            _ = obj.thermo_boundaries.pop(self.name)
+
+    def _build_table(self, obj, value):
+        """Set the chemical potentials and update table."""
+        for spec, count in Counter(map(get_species, value.keys())).items():
+            if count > 1:
+                raise ValueError(
+                    f"{count} values of the chemical potential for the same "
+                    f"species {spec} were provided.\n Make sure the dictionary "
+                    "you are using has only string keys or only Species "
+                    "objects as keys."
+                )
+        value = {get_species(k): v for k, v in value.items() if k in obj.species}
+        if set(value.keys()) != set(obj.species):
+            raise ValueError(
+                "Chemical potentials given are missing species. "
+                "Values must be given for each of the following:"
+                f" {obj.species}"
+            )
+        num_cols = max(max(sl.encoding) for sl in obj.sublattices) + 1
+
+        # Sublattice can only be initialized as default, or splitted from default.
+        table = np.zeros((obj.num_sites, num_cols))
+        for sublatt in obj.active_sublattices:
+            ordered_pots = [value[sp] for sp in sublatt.site_space]
+            table[sublatt.sites[:, None], sublatt.encoding] = ordered_pots
+        return table
+
+
+class Ensemble(MSONable):
+    """Thermodynamic ensemble class.
+
+    Attributes:
+        thermo_boundaries (dict):
+            Dictionary with corresponding thermodynamic boundaries, i.e.
+            chemical potentials or fugacity fractions. This is kept only for
+            descriptive purposes.
+    """
+
+    chemical_potentials = ChemicalPotentials()
+
+    def __init__(self, processor, sublattices=None):
+        """Initialize class instance.
+
+        Args:
+            processor (Processor):
+                a processor that can compute the change in property given
+                a set of flips.
+            sublattices (list of Sublattice): optional
+                list of Sublattice objects representing sites in the processor
+                supercell with same site spaces.
+        """
+        if sublattices is None:
+            sublattices = processor.get_sublattices()
+        self.thermo_boundaries = {}  # not pretty way to save general info
+        self._params = processor.coefs  # natural parameters
[…]
