+++ smol/moca/__init__.py
@@ -8,6 +8,7 @@ Monte Carlo simulations using Cluster Expansion Hamiltonians.
 from smol.moca.sampler.container import SampleContainer
 
 from .ensemble.canonical import CanonicalEnsemble
+from .ensemble.ensemble import Ensemble
 from .ensemble.semigrand import SemiGrandEnsemble
 from .processor.composite import CompositeProcessor
 from .processor.ewald import EwaldProcessor
@@ -18,6 +19,7 @@ __all__ = [
     "ClusterExpansionProcessor",
     "EwaldProcessor",
     "CompositeProcessor",
+    "Ensemble",
     "CanonicalEnsemble",
     "SemiGrandEnsemble",
     "Sampler",
