+++ smol/moca/sampler/sampler.py
@@ -64,8 +64,8 @@ class Sampler:
             ensemble (Ensemble):
                 an Ensemble class to obtain sample probabilities from.
             step_type (str): optional
-                type of step to run MCMC with. If not given the default is the
-                first entry in the Ensemble.valid_mcmc_steps.
+                type of step to run MCMC with. If not given the default depends on
+                whether chemical potentials are defined.
             *args:
                 positional arguments to pass to the MCKernel constructor.
                 More often than not you want to specify the temperature!
@@ -86,12 +86,14 @@ class Sampler:
             Sampler
         """
         if step_type is None:
-            step_type = ensemble.valid_mcmc_steps[0]
-        elif step_type not in ensemble.valid_mcmc_steps:
-            raise ValueError(
-                f"Step type {step_type} can not be used for sampling a "
-                f"{type(ensemble)}."
-            )
+            if (
+                hasattr(ensemble, "chemical_potentials")
+                and ensemble.chemical_potentials is not None
+            ):
+                step_type = "flip"
+            else:
+                step_type = "swap"
+
         if kernel_type is None:
             kernel_type = "Metropolis"
 
