### 2022-05-25 ff5dec4
fix passing seed to Sampler.from_ensemble
