+++ CHANGES.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Use this section to keep track of changes in the works.
 ### Added
 ### Changed
+### Removed
+### Deprecated
+
+# [v0.0.2](https://github.com/CederGroupHub/smol/releases/tag/v0.0.2) (2022-05-22)
 ### Fixed
 * version dunder with pypi project rename.
 * use of np.random.default_rng for reproducibility. #206 (lbluque)
-### Removed
-### Deprecated
+* Fix passing seed explicitly in Sampler.from_ensemble
 
 # [v0.0.1](https://github.com/CederGroupHub/smol/releases/tag/v0.0.1) (2022-04-26)
 ### Added
