### 2022-05-25 1c1828e
fix chemical potential manager
