+++ smol/moca/sampler/container.py
@@ -202,14 +202,16 @@ class SampleContainer(MSONable):
         """Get the variance of feature vector elements."""
         return self.get_feature_vectors(discard, thin_by, flat).var(axis=0)
 
-    def get_orbit_factors(self, function_orbit_ids, discard=0,
-                          thin_by=1, flat=True):
+    def get_orbit_factors(self, function_orbit_ids, discard=0, thin_by=1, flat=True):
         """Get the orbit factor vectors for samples."""
         vals = self.natural_parameters * self.get_feature_vectors(
-            discard=discard, thin_by=thin_by, flat=flat)
+            discard=discard, thin_by=thin_by, flat=flat
+        )
         orbit_factors = np.array(
-            [np.sum(vals[function_orbit_ids == i])
-                for i in range(len(self.natural_parameters))]
+            [
+                np.sum(vals[function_orbit_ids == i])
+                for i in range(len(self.natural_parameters))
+            ]
         )
         return orbit_factors
 
