+++ smol/moca/processor/__init__.py
@@ -8,8 +8,12 @@ to run Monte Carlo sampling.
 """
 
 from .composite import CompositeProcessor
-from .expansion import ClusterExpansionProcessor, ClusterDecompositionProcessor
 from .ewald import EwaldProcessor
+from .expansion import ClusterDecompositionProcessor, ClusterExpansionProcessor
 
-__all__ = ['CompositeProcessor', 'ClusterExpansionProcessor',
-           'ClusterDecompositionProcessor', 'EwaldProcessor']
+__all__ = [
+    "CompositeProcessor",
+    "ClusterExpansionProcessor",
+    "ClusterDecompositionProcessor",
+    "EwaldProcessor",
+]
