+++ tests/test_cofe/test_clusterspace.py
@@ -624,15 +624,20 @@ def test_periodicity(single_subspace):
     npt.assert_allclose(a, b)
 
 
-def _encode_occu(occu, bits):
-    return np.array([bit.index(sp) for sp, bit in zip(occu, bits)])
+def test_site_basis_rotation(cluster_subspace):
+    cs = cluster_subspace.copy()
+    if not cs.basis_orthogonal:
+        with pytest.raises(RuntimeError):
+            cs.rotate_site_basis(1, np.pi / 4)
+        cs.change_site_bases("sinusoid", orthonormal=True)
 
+    with pytest.raises(ValueError):
+        cs.rotate_site_basis(len(cs.orbits_by_size[1]) + 2, np.pi)
 
-def test_site_basis_rotation(self):
-    cs = self.cs.copy()
-    cs.change_site_bases("sinusoid", orthonormal=True)
     cs1 = cs.copy()
+    print(cs.site_rotation_matrix)
     cs1.rotate_site_basis(1, np.pi / 4)
+    print(cs1.site_rotation_matrix)
     for i in range(5):
         structure = gen_random_structure(cs.structure)
         for j in range(5):
@@ -664,27 +669,28 @@ def test_site_basis_rotation(self):
                 ]
             )
             # test ECI invariance
-            self.assertTrue(np.allclose(norms, norms1))
+            assert np.allclose(norms, norms1)
             # test correlation vector invariance
-            self.assertTrue(
+            assert (
                 np.allclose(
                     cs1.site_rotation_matrix @ cs.corr_from_structure(structure),
                     cs1.corr_from_structure(structure),
                 )
             )
             # test values are the same
-            self.assertTrue(
+            assert (
                 np.isclose(
                     np.dot(cs.corr_from_structure(structure), coefs),
                     np.dot(cs1.corr_from_structure(structure), coefs1),
                 )
             )
-            # test rotation matrix is invertible det = 1
-            self.assertTrue(np.isclose(np.linalg.det(cs1.site_rotation_matrix), 1))
+            # test rotation matrix is invertible and abs(det) = 1
+            # binary site rotations may lead to det = -1 ?
+            assert np.isclose(abs(np.linalg.det(cs1.site_rotation_matrix)), 1)
 
-    @staticmethod
-    def _encode_occu(occu, bits):
-        return np.array([bit.index(sp) for sp, bit in zip(occu, bits)])
+
+def _encode_occu(occu, bits):
+    return np.array([bit.index(sp) for sp, bit in zip(occu, bits)])
 
 
 def test_vs_CASM_pairs(single_structure):
