### 2022-05-26 d927c23
remove __init__ from ChemPotManager
