### 2022-05-26 a15eb2a
_build_table staticmethod
