### 2022-05-26 38e1522
reset chemical_potentials in split_sublattices
