### 2022-05-26 325a5f3
update tests for single ensemble
