+++ tests/test_moca/test_ensemble.py
@@ -6,36 +6,31 @@ import pytest
 
 from smol.cofe import ClusterExpansion, RegressionData
 from smol.moca import (
-    CanonicalEnsemble,
     ClusterExpansionProcessor,
     CompositeProcessor,
+    Ensemble,
     EwaldProcessor,
-    SemiGrandEnsemble,
 )
 from tests.utils import assert_msonable, gen_random_occupancy
-
-ensembles = [CanonicalEnsemble, SemiGrandEnsemble]
 
 
 @pytest.fixture
 def canonical_ensemble(composite_processor):
-    return CanonicalEnsemble(composite_processor)
+    return Ensemble(composite_processor)
 
 
 @pytest.fixture
-def mugrand_ensemble(composite_processor):
+def semigrand_ensemble(composite_processor):
     species = {
         sp for space in composite_processor.active_site_spaces for sp in space.keys()
     }
     chemical_potentials = {sp: 0.3 for sp in species}
-    return SemiGrandEnsemble(
-        composite_processor, chemical_potentials=chemical_potentials
-    )
+    return Ensemble(composite_processor, chemical_potentials=chemical_potentials)
 
 
 # Test with composite processors to cover more ground
-@pytest.mark.parametrize("ensemble_cls", ensembles)
-def test_from_cluster_expansion(cluster_subspace_ewald, ensemble_cls, rng):
+@pytest.mark.parametrize("ensemble_type", ["canonical", "semigrand"])
+def test_from_cluster_expansion(cluster_subspace_ewald, ensemble_type, rng):
     coefs = rng.random(cluster_subspace_ewald.num_corr_functions + 1)
     scmatrix = 3 * np.eye(3)
     proc = CompositeProcessor(cluster_subspace_ewald, scmatrix)
@@ -61,13 +56,13 @@ def test_from_cluster_expansion(cluster_subspace_ewald, ensemble_cls, rng):
     )
     expansion = ClusterExpansion(cluster_subspace_ewald, coefs, reg_data)
 
-    if ensemble_cls is SemiGrandEnsemble:
+    if ensemble_type == "semigrand":
         species = {sp for space in proc.active_site_spaces for sp in space.keys()}
         chemical_potentials = {sp: 0.3 for sp in species}
         kwargs = {"chemical_potentials": chemical_potentials}
     else:
         kwargs = {}
-    ensemble = ensemble_cls.from_cluster_expansion(
+    ensemble = Ensemble.from_cluster_expansion(
         expansion, supercell_matrix=scmatrix, **kwargs
     )
     npt.assert_array_equal(
@@ -115,9 +110,9 @@ def test_split_ensemble(ensemble):
         S = len(sublattice.species)
         old_sublattices = deepcopy(ensemble.sublattices)
         old_species = deepcopy(ensemble.species)
-        if isinstance(ensemble, SemiGrandEnsemble):
+        if ensemble.chemical_potentials is not None:
             old_chemical_potentials = deepcopy(ensemble.chemical_potentials)
-            old_mu_table = deepcopy(ensemble._mu_table)
+            old_mu_table = deepcopy(ensemble._chemical_potentials["table"])
         split_encodings = [sublattice.encoding[: S // 2], sublattice.encoding[S // 2 :]]
         ensemble.split_sublattice_by_species(sl_id, occu, split_encodings)
         assert len(ensemble.sublattices) == len(old_sublattices) + 1
@@ -154,9 +149,9 @@ def test_split_ensemble(ensemble):
                 )
                 assert len(new1.encoding) == len(new1.species)
                 assert len(new2.encoding) == len(new2.species)
-        if isinstance(ensemble, SemiGrandEnsemble):
[…]
