### 2022-05-26 791f5be
add deprecation warning to old ensembles
