### 2022-05-26 331329e
fix processor tests
