### 2022-05-26 4d09a0f
add tests for chempotmanager
