### 2022-05-27 #4356 @jarvist
@TomTranter if you would still be happy to review, that would be much appreciated! @kitmccoll has cluster expansion experience, and it's nicely complementary to have someone naive to the science as other reviewer, in case there's too much lingo / jargon.
