### 2022-05-27 62e17f4
add LICENSE to index.rst
