+++ docs/src/index.rst
@@ -44,7 +44,7 @@ related methods. Although **smol** is intentionally lightweight---in terms of
 dependencies and built-in functionality---it has a modular design that closely
 follows underlying mathematical formalism and provides useful abstractions to
 easily extend existing methods or implement and test new ones. Finally,
-although conceived mainly for method development, **smol** can (and is being)
+although initially conceived for method development, **smol** can (and is being)
 used in production for materials science reasearch applications.
 
 Functionality
