+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name  =  statmech-on-lattices
+name  =  smol
 version  =  attr: smol.__version__
 author  =  Luis Barroso-Luque
 author_email  =  lbluque@berkeley.edu
