+++ .github/workflows/build.yml
@@ -9,7 +9,9 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-20.04, windows-2019, macos-10.15]
+        os: [ubuntu-latest, windows-latest, macos-latest]
+    env:
+      CIBW_SKIP: pp*  # skip pypy wheels
 
     steps:
       - uses: actions/checkout@v2
