+++ smol/moca/sampler/mcusher.py
@@ -24,7 +24,7 @@ from ..utils.occu_utils import (get_dim_ids_by_sublattice,
                                 occu_to_species_n,
                                 delta_n_from_step)
 from ..utils.math_utils import (choose_section_from_partition, gcd_list,
-                                comb, flip_weights_mask)
+                                comb, flip_weights_mask, NUM_TOL)
 
 
 class MCUsher(ABC):
@@ -200,7 +200,8 @@ class Tableflip(MCUsher):
     write.
     """
 
-    def __init__(self, sublattices, flip_table=None,
+    def __init__(self, sublattices, rng=None,
+                 flip_table=None,
                  charge_balanced=True,
                  other_constraints=None,
                  optimize_basis=False,
@@ -213,6 +214,10 @@ class Tableflip(MCUsher):
                 list of Sublattices to propose steps for. Must
                 be all sublattices, including active and inactive
                 ones.
+            rng (np.Generator): optional
+                The given PRNG must be the same instance as that used by the
+                kernel and any bias terms, otherwise reproducibility will be
+                compromised.
             flip_table(2D np.ndarray[int]), optional:
                 Flip vectors in "n" format, as generated by a CompSpace.
                 If not given, we will initialize a CompSpace from
@@ -264,7 +269,7 @@ class Tableflip(MCUsher):
         sub-lattices is not meaningful. We will always assign
         sub-lattice probabilties as equal.
         """
-        super(Tableflip, self).__init__(sublattices)
+        super(Tableflip, self).__init__(sublattices, rng=rng)
         self.bits = [sl.species for sl in self.sublattices]
         self.dim_ids = get_dim_ids_by_sublattice(self.bits)
         self.sl_sizes = np.array([len(sl.sites) for sl in self.sublattices],
@@ -281,6 +286,7 @@ class Tableflip(MCUsher):
                                     other_constraints=other_constraints,
                                     optimize_basis=optimize_basis,
                                     table_ergodic=table_ergodic)
+
         if flip_table is not None:
             self.flip_table = np.array(flip_table, dtype=int)
         else:
@@ -307,6 +313,8 @@ class Tableflip(MCUsher):
         self._swapper = Swap(self.sublattices)
         self._dim_ids_table = get_dim_ids_table(self.sublattices,
                                                 active_only=True)
+        self._dim_ids_full = get_dim_ids_table(self.sublattices,
+                                               active_only=False)
 
     def propose_step(self, occupancy):
         """Propose a single random flip step.
@@ -332,25 +340,42 @@ class Tableflip(MCUsher):
         Returns:
              list(tuple): list of tuples each with (idex, code).
         """
-        rng = np.random.default_rng()
+        rng = self._rng
         if rng.random() < self.swap_weight:
             return self._swapper.propose_step(occupancy)
 
-        # We shall only flip active sites.
+        # We shall only flip active sites. And only active sites stated.
         species_list = occu_to_species_list(occupancy, self.d,
                                             self._dim_ids_table)
         species_n = [len(sites) for sites in species_list]
-        mask = flip_weights_mask(self.flip_table,
-                                 species_n, self.max_n).astype(int)
+        species_list_full = occu_to_species_list(occupancy,
+                                                 self.d,
+                                                 self._dim_ids_full)
+        species_n_full = [len(sites) for sites in species_list_full]
+
+        if not np.allclose(self._compspace.A @ np.array(species_n_full),
+                           self._compspace.b * self.sc_size):
+            warnings.warn("Current occupancy violates compspace constraints! "
+                          "Are you initializing trace?")
+            mask = np.zeros(2 * len(self.flip_table), dtype=int)
+        else:
+            mask = flip_weights_mask(self.flip_table,
[…]
