+++ docs/src/api_reference/moca/ensembles.base.rst
@@ -1,8 +1,0 @@
-=============
-Base Ensemble
-=============
-
-.. automodule:: smol.moca.ensemble.base
-   :members:
-   :undoc-members:
-   :show-inheritance:
