+++ smol/capp/sqs.py
@@ -1,0 +1,6 @@
+"""Generate special structures"""
+
+
+def special_structure_anneal(processor, target_features=None):
+    """special structures."""
+    pass
