+++ docs/src/developer_guide/index.rst
@@ -5,10 +5,11 @@ Developing
 .. note::
    This page is still under construction...
 
-All classes and functions are designed for easy extension and implementation
-of new metholody.
+The following sections describe the design of the package in detail as well as various contributing guidelines to make
+writting and contributing new functionality much more straightforward.
 
 .. toctree::
-   :maxdepth: 1
-   guidelines
+   :maxdepth: 3
+
+   guide
    design
