+++ smol/moca/sampler/container.py
@@ -14,7 +14,7 @@ from collections import defaultdict
 
 import numpy as np
 from monty.dev import requires
-from monty.json import MSONable, jsanitize
+from monty.json import MSONable, MontyEncoder, MontyDecoder, jsanitize
 
 from smol.moca.sampler.kernel import Trace
 from smol.moca.sublattice import Sublattice
@@ -407,7 +407,7 @@ class SampleContainer(MSONable):
         backend["natural_parameters"].attrs[
             "num_energy_coefs"
         ] = self._num_energy_coefs  # noqa
-        backend.create_dataset("sampling_metadata", data=json.dumps(self.metadata))
+        backend.create_dataset("sampling_metadata", data=json.dumps(jsanitize(self.metadata)))
         trace_grp = backend.create_group("trace")
         for name, value in self._trace.items():
             trace_grp.create_dataset(
@@ -447,7 +447,7 @@ class SampleContainer(MSONable):
             "@class": self.__class__.__name__,
             "sublattices": [s.as_dict() for s in self.sublattices],
             "natural_parameters": jsanitize(self.natural_parameters),
-            "metadata": self.metadata,
+            "metadata": jsanitize(self.metadata),
             "num_energy_coefs": self._num_energy_coefs,
             "total_mc_steps": self._total_steps,
             "nsamples": self._nsamples,
@@ -537,7 +537,8 @@ class SampleContainer(MSONable):
                     "num_energy_coefs"
                 ],  # noqa
                 sample_trace=trace,
-                sampling_metadata=json.loads(f["sampling_metadata"][()]),
+                sampling_metadata=json.loads(
+                    f["sampling_metadata"][()], cls=MontyDecoder),
             )
             container._nsamples = nsamples
             container._total_steps = f["trace"].attrs["total_mc_steps"]
