### 2022-06-02 846831e
change pip install smol in docs
