### 2022-06-02 0a63391
add svg diagrams to docs
