+++ docs/src/developer_guide/index.rst
@@ -2,11 +2,10 @@
 Developing
 ==========
 
-.. note::
-   This page is still under construction...
-
-The following sections describe the design of the package in detail as well as various contributing guidelines to make
-writting and contributing new functionality much more straightforward.
+The following sections list contributing guidelines as well a detailed description of
+design of the package that may help to understand much better how and where to implement
+new features. This information should help make writting and contributing new
+functionality much more straightforward.
 
 .. toctree::
    :maxdepth: 3
