+++ docs/src/getting_started.rst
@@ -31,8 +31,7 @@ Start by creating a disordered primitive structure.
 
     >>> from pymatgen.core.structure import Structure, Lattice
     >>> species = {"Au": 0.5, "Cu": 0.5}
-    >>> prim = Structure.from_spacegroup(
-            "Fm-3m", Lattice.cubic(3.6), [species], [[0, 0, 0]])
+    >>> prim = Structure.from_spacegroup("Fm-3m", Lattice.cubic(3.6), [species], [[0, 0, 0]])
 
 Now create a cluster subspace for that structure including pair, triplet and
 quadruplet clusters up to given cluster diameter cutoffs.
@@ -72,8 +71,7 @@ fit.
 
     >>> from sklearn.linear_model import LinearRegression
     >>> reg = LinearRegression(fit_intercept=False)
-    >>> reg.fit(wrangler.feature_matrix,
-                wrangler.get_property_vector("energy"))
+    >>> reg.fit(wrangler.feature_matrix, wrangler.get_property_vector("energy"))
 
 Finally, create a cluster expansion for prediction of new structures and
 eventual Monte Carlo sampling. We recommed saving the details used to fit the
@@ -85,9 +83,9 @@ expansion for future reproducibility (although this is not strictly necessary).
     >>> reg_data = RegressionData.from_sklearn(
             estimator=reg,
             feature_matrix=wrangler.feature_matrix,
-            property_vector=wrangler.get_property_vector('energy'))
-    >>> expansion = ClusterExpansion(
-            subspace, coefficients=reg.coef_, regression_data=reg_data)
+            property_vector=wrangler.get_property_vector('energy')
+        )
+    >>> expansion = ClusterExpansion(subspace, coefficients=reg.coef_, regression_data=reg_data)
 
 Creating an ensemble for Monte Carlo Sampling
 ---------------------------------------------
@@ -97,10 +95,9 @@ to define the sampling domain.
 
 .. nbplot::
 
-    >>> from smol.moca import CanonicalEnsemble
+    >>> from smol.moca import Ensemble
     >>> sc_matrix = [[5, 0, 0], [0, 5, 0], [0, 0, 5]]
-    >>> ensemble = CanonicalEnsemble.from_cluster_expansion(
-            expansion, supercell_matrix=sc_matrix)
+    >>> ensemble = Ensemble.from_cluster_expansion(expansion, supercell_matrix=sc_matrix)
 
 Running Monte Carlo sampling
 ----------------------------
@@ -110,8 +107,7 @@ object.
 .. nbplot::
 
     >>> from smol.moca import Sampler
-    >>> sampler = Sampler.from_ensemble(
-            ensemble, temperature=500)
+    >>> sampler = Sampler.from_ensemble(ensemble, temperature=1000)
 
 In order to begin an MC simulation, an initial configuration must be provided.
 In this case we use pymatgen's functionality to provide an ordered structure
@@ -119,8 +115,7 @@ given a disordered one.
 
 .. nbplot::
 
-    >>> from pymatgen.transformations.standard_transformations import \
-            OrderDisorderedStructureTransformation
+    >>> from pymatgen.transformations.standard_transformations import OrderDisorderedStructureTransformation
     >>> transformation = OrderDisorderedStructureTransformation()
     >>> structure = expansion.cluster_subspace.structure.copy()
     >>> structure.make_supercell(sc_matrix)
@@ -132,7 +127,7 @@ to run MC sampling interations.
 .. nbplot::
 
     >>> init_occu = ensemble.processor.occupancy_from_structure(structure)
-    >>> sampler.run(1000000, initial_occupancy=init_occu)
[…]
