+++ docs/src/developer_guide/design.rst
@@ -4,7 +4,7 @@
 Package Design
 ==============
 
-Overview & mission
+Overview & Mission
 ==================
 
 **smol** is intentionally designed to be easy to use, install and extend. In order to
@@ -28,6 +28,17 @@ dependent properties of inorganic materials.
 Module Design
 =============
 
+**smol** is organized into two main submodules, :mod:`smol.cofe` and
+:mod:`smol.moca`:
+
+- :mod:`smol.cofe` implements the classes and functions to represent
+  site spaces, configuration spaces and function subspaces over them, and the tools
+  necessary to fit lattice models, in the spirit of *cluster expansions* and the
+  *structure inversion method*.
+- :mod:`smol.moca` has implementations of classes to carry out Monte Carlo sampling of
+  ensembles associated with a fitted lattice model Hamiltonian, as well as simple
+  empirical models such as a point Coulomb potential.
+
 Below is a detailed description of the classes and their relationships in the main
 modules. Before delving into these we recommend going over the simpler overview of the
 package in the :ref:`user_guide` page.
@@ -41,15 +52,135 @@ Ownership relationships between distinct classes (i.e. when a class has another 
 as an attribute) are depicted as dark blue colored arrows. The arrow points from the
 class which is an attribute to the class that holds the former as an attributed.
 
-smol.cofe
----------
+:mod:`smol.cofe`
+----------------
+
+The diagram below summarizes the overall design of the :mod:`smol.cofe` module. The main
+classes that users directly instantiate and interact with are the
+:class:`ClusterSubspace`, :class:`StructureWrangler`, and :class:`ClusterExpansion` as
+detailed in the :ref:`user_guide`. All main classes are derived from :mod:`monty`
+:class:`MSONable` and thus can be serialized and saved as JSON objects.
+In addition, several other classes are defined to enable the various features and
+functionality in the module.
 
 .. image:: ../../_static/smol_cofe.svg
 
-smol.moca
----------
+Following the diagram above, the main purpose of the classes depicted is,
+
+* :class:`ClusterSubspace` (and its derived class :class:`PottsSubspace`) is the main
+  work horse to construct and represent configuration spaces and the function spaces
+  over them. Its main purpose is to compute *correlation function* values given an
+  ordered structure. In order to do so, it holds a list of :class:`Orbit` objects
+  generated for a particular disordered :mod:`pymatgen` :class:`Structure`. Additionally
+  an *external term* representing a simple empricial potential model can also be
+  included. Currently only an :class:`EwaldTerm`, to represent a point electrostatic
+  potential is available.
+
+  * An :class:`Orbit` represents both the set of symmetrically equivalent clusters, as
+    well as the set of *product functions* that act over the configurations of those
+    clusters. An orbit holds a base :class:`Cluster` and a list of :ref:`SiteBasis`
+    associated with each site in the cluster.
+
+    * A :class:`Cluster` is a collection of sites (derived from :class:`SiteCollection`
+      in :mod:`pymatgen`). The two key concepts of a :class:`Cluster` is that compared
+      to a :class:`Structure` they do hold periodic sites, and compared to a
+      :class:`Molecule` the hold a lattice (of the structure they are associated with).
+    * A :class:`SiteBasis` represents the basis set that spans the function space of
+      the configurations for a single site. Several types of basis sets are included
+      and implementing new ones is relatively straighforward. A :class:`SiteBasis` holds
+      a :class:`SiteSpace` which represents its single site configuration space.
+
+      * :class:`SiteSpace` represents the possible configurations of a given site.
+        It essentially holds a :mod:`pymatgen` :class:`Composition`, with the addition
+        of an explicit :class:`Vacancy` species, for compositions that do not sum to 1.
+
[…]
