+++ smol/moca/sampler/mcusher.py
@@ -50,7 +50,7 @@ class MCUsher(ABC):
             )
         elif len(sublattice_probabilities) != len(self.active_sublattices):
             raise AttributeError(
-                "Sublattice probabilites needs to be the " "same length as sublattices."
+                "Sublattice probabilites needs to be the same length as sublattices."
             )
         elif sum(sublattice_probabilities) != 1:
             raise ValueError("Sublattice probabilites must sum to one.")
@@ -156,6 +156,94 @@ class Swap(MCUsher):
             # inefficient, maybe re-call method? infinite recursion problem
             swap = []
         return swap
+
+
+class Composite(MCUsher):
+    """A composite usher for chaining different step types.
+
+    This can be used to mix step types for all sublattices or to create hybrid
+    ensembles where different sublattices have different allowed steps.
+    """
+
+    def __init__(
+        self,
+        sublattices,
+        sublattice_probabilities=None,
+        mcushers=None,
+        weights=None,
+        rng=None,
+    ):
+        """Initialize the composite usher.
+
+        Args:
+            sublattices (list of Sublattice):
+                list of active Sublattices to propose steps for. Active
+                sublattices are those that include sites with configuration
+                degrees of freedom DOFs, only occupancy on active sub-lattices' active sites
+                are allowed to change.
+            sublattice_probabilities (list of float): optional
+                list of probabilities to pick a site from specific active
+                sublattices.
+            mcushers (list of MCUsher):
+                A list of mcushers to add to the composite.
+            weights (list of float):
+                A list of the weights associated with each mcusher passed. With this
+                the corresponding probabilities for each mcusher to be picked are
+                generated. Must be in the same order as the mcusher list.
+            rng (np.Generator): optional
+                The given PRNG must be the same instance as that used by the kernel and
+                any bias terms, otherwise reproducibility will be compromised.
+        """
+        super().__init__(sublattices, sublattice_probabilities, rng=rng)
+        self._mcushers = []
+        self._weights = []
+        self._p = []
+
+        if mcushers is not None:
+            if weights is None:
+                weights = len(mcushers) * [
+                    1,
+                ]
+            for weight, usher in zip(weights, mcushers):
+                if isinstance(usher, str):
+                    usher = mcusher_factory(
+                        usher,
+                        self.sublattices,
+                        sublattice_probabilities=self.sublattice_probabilities,
+                    )
+                self.add_mcusher(usher, weight)
+
+    @property
+    def mcushers(self):
+        """Get the list of mcushers."""
+        return self._mcushers
+
+    @property
+    def weight(self):
+        """Get the weights associated with each mcusher."""
+        return self._weights
+
+    def add_mcusher(self, mcusher, weight=1):
+        """Add an MCUsher to the Composite usher.
+
+        Args:
+            mcusher (MCUsher):
+                The MCUsher to add
+            weight (float):
+                The weight associated with the mcusher being added
+        """
+        self._mcushers.append(mcusher)
+        self._update_p(weight)
+
+    def _update_p(self, weight):
+        """Update the probabilities for each mcuhser based on a new weight."""
+        self._weights.append(weight)
+        total = sum(self._weights)
+        self._p = [weight / total for weight in self._weights]
+
+    def propose_step(self, occupancy):
+        """Propose a step given an occupancy."""
+        return self._rng.choice(self._mcushers, p=self._p).propose_step(occupancy)
 
 
 def mcusher_factory(usher_type, sublattices, *args, **kwargs):
