### 2022-06-03 0915124
Update build.yml
set OS versions as in docs
