+++ .github/workflows/release.yml
@@ -64,28 +64,64 @@ jobs:
           draft: false
           prerelease: false
 
-  release-pypi:
+  build_wheels:
+    name: Build wheels on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
 
+    needs:
+      - release-github
+    strategy:
+      matrix:
+        os: [ubuntu-20.04, windows-2019, macos-10.15]
+    env:
+      CIBW_SKIP: pp*  # skip pypy wheels
+
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: Build wheels
+        uses: pypa/cibuildwheel@v2.6.0
+
+      - uses: actions/upload-artifact@v3
+        with:
+          path: wheelhouse/*.whl
+
+  build_sdist:
+    name: Build source distribution
     runs-on: ubuntu-latest
+
     needs:
       - release-github
 
     steps:
       - uses: actions/checkout@v3
-
-      - uses: actions/setup-python@v3
         with:
-           python-version: 3.9
+          fetch-depth: 0 # Optional, use if you use setuptools_scm
 
-      - name: Build
-        run: |
-          pip install setuptools setuptools_scm numpy build
-          python -m build --sdist
+      - name: Build sdist
+        run: pipx run build --sdist
 
-      - name: Publish package
-        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-        uses: pypa/gh-action-pypi-publish@release/v1
+      - uses: actions/upload-artifact@v3
         with:
-          verbose: true
-          user: __token__
-          password: ${{ secrets.PYPI_API_TOKEN }}
+          path: dist/*.tar.gz
+
+  release-pypi:
+    runs-on: ubuntu-latest
+    needs: [build_wheels, build_sdist]
+
+    steps:
+      steps:
+        - uses: actions/download-artifact@v3
+          with:
+            name: artifact
+            path: dist
+
+        - name: Publish package
+          if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+          uses: pypa/gh-action-pypi-publish@release/v1
+          with:
+            verbose: true
+            user: __token__
+            password: ${{ secrets.PYPI_API_TOKEN }}
