+++ .github/workflows/release.yml
@@ -112,16 +112,15 @@ jobs:
     needs: [build_wheels, build_sdist]
 
     steps:
-      steps:
-        - uses: actions/download-artifact@v3
-          with:
-            name: artifact
-            path: dist
+      - uses: actions/download-artifact@v3
+        with:
+          name: artifact
+          path: dist
 
-        - name: Publish package
-          if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-          uses: pypa/gh-action-pypi-publish@release/v1
-          with:
-            verbose: true
-            user: __token__
-            password: ${{ secrets.PYPI_API_TOKEN }}
+      - name: Publish package
+        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          verbose: true
+          user: __token__
+          password: ${{ secrets.PYPI_API_TOKEN }}
