### 2022-06-04 833de1c
allow streaming in annealing (+ a few fixes)
