### 2022-06-04 813ec94
add comment to anneal docstring
