### 2022-06-05 c20d373
fix clearing samples in run
