+++ CHANGES.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Use this section to keep track of changes in the works.
 ### Added
+* Allow streaming to h5 in simulated annealing. #216 @lbluque
 ### Changed
 ### Fixed
 ### Removed
