### 2022-06-05 1b4f57d
use saved rng in WL kernel
