+++ smol/moca/sampler/kernel.py
@@ -436,22 +436,22 @@ class WangLandau(MCKernel):
     """
 
     valid_mcushers = ALL_MCUSHERS
-    valid_bias = ALL_BIAS
+    valid_bias = None
 
     def __init__(
         self,
         ensemble,
         step_type,
         bin_size,
-        min_energy,
-        max_energy,
         flatness=0.8,
         mod_factor=1.0,
         check_period=1000,
-        fixed_window=False,
+        min_energy=None,
+        max_energy=None,
         mod_update=None,
         seed=None,
         nwalkers=1,
+        **kwargs,
     ):
         """Initialize a WangLandau Kernel.
 
@@ -462,11 +462,6 @@ class WangLandau(MCKernel):
                 An MC step type corresponding to an MCUsher. See valid_mcushers
             bin_size (float):
                 The energy bin size to determine different states.
-            min_energy (float):
-                The minimum energy to sample. Energy value should be given per
-                supercell (i.e. same order as what will be sampled).
-            max_energy (float):
-                The maximum energy to sample.
             flatness (float): optional
                 The flatness factor used when checking histogram flatness.
                 Must be between 0 and 1.
@@ -476,9 +471,11 @@ class WangLandau(MCKernel):
             check_period (int): optional
                 The period in number of steps for the histogram flatness to be
                 checked.
-            fixed_window (bool): optional
-                Whether to accept energies below the minimum and above the
-                maximum energy provided. Default is False
+            min_energy (float):
+                The minimum energy to sample. Energy value should be given per
+                supercell (i.e. same order as what will be sampled).
+            max_energy (float):
+                The maximum energy to sample. Same as above.
             mod_update (Callable): optional
                 A function used to update the fill factor when the histogram
                 satisfies the flatness criteria. The function is used to update
@@ -489,19 +486,14 @@ class WangLandau(MCKernel):
             nwalkers (int): optional
                 Number of walkers/chains to sampler. Default is 1.
         """
-        if min_energy > max_energy:
-            raise ValueError("min_energy can not be larger than max_energy.")
+        if min_energy is not None and max_energy is not None:
+            if min_energy > max_energy:
+                raise ValueError("min_energy can not be larger than max_energy.")
         elif mod_factor <= 0:
             raise ValueError("mod_factor must be greater than 0.")
-        elif fixed_window is True:
-            raise NotImplementedError(
-                "fixed_window=True is not implemented."
-                " If you need this, consider doing a PR."
-            )
 
         self.flatness = flatness
         self.check_period = check_period
-        self.fixed_window = fixed_window
         self._mfactors = np.array(
             nwalkers
             * [
@@ -510,35 +502,34 @@ class WangLandau(MCKernel):
         )
         self._window = (min_energy, max_energy)
         self._bin_size = bin_size
-        self._update_fun = mod_update if mod_update is not None else lambda f: f / 2.0
+        self._mod_update = mod_update if mod_update is not None else lambda f: f / 2.0
 
         # default dict of arrays with [entropy, histogram]
         # keys are generated by _get_key method
-        nbins = int((max_energy - min_energy) // bin_size)
+        # nbins = int((max_energy - min_energy) // bin_size)
         self._aux_states = defaultdict(
             lambda: np.array(
                 nwalkers
                 * [
                     [0.0, 0.0],
                 ]
-            ),
-            {
-                bin_num: np.array(
-                    nwalkers
-                    * [
-                        [0.0, 0.0],
-                    ]
-                )
[…]
