### 2022-06-06 99ee924
fix recording traces for multiple walkers (need to clean up this interface!)
