+++ smol/moca/sampler/sampler.py
@@ -197,7 +197,10 @@ class Sampler:
                         map(self._kernel.single_step, occupancies)
                     ):
                         for name, value in strace.items():
-                            setattr(trace, name, value)
+                            val = getattr(trace, name)
+                            val[i] = value
+                            # this will mess up recording values for > 1 walkers
+                            # setattr(trace, name, value)
                         if strace.accepted:
                             for name, delta_val in strace.delta_trace.items():
                                 val = getattr(trace, name)
