### 2022-06-06 b6d9b4b
fix computing intial trace
