### 2022-06-07 31e27fd
fix resetting aux states in WL
