+++ smol/moca/sampler/kernel.py
@@ -7,9 +7,7 @@ to generate states for sampling an MCMC chain.
 __author__ = "Luis Barroso-Luque"
 
 from abc import ABC, abstractmethod
-from collections import defaultdict
 from math import log
-from operator import itemgetter
 from types import SimpleNamespace
 
 import numpy as np
@@ -103,6 +101,7 @@ class StepTrace(Trace):
 # TODO make it easier to have multiple walkers, either have the sampler have
 #  a list of kernel copies or make a multi-kernel class that simply holds
 #  the copies but ow behaves the same, that will really simplify writing kernels!
+#  TODO it is so tedious and unessecary to have multiple walkers for the WL Kernel !
 
 
 class MCKernel(ABC):
@@ -443,11 +442,11 @@ class WangLandau(MCKernel):
         ensemble,
         step_type,
         bin_size,
+        min_energy,
+        max_energy,
         flatness=0.8,
         mod_factor=1.0,
         check_period=1000,
-        min_energy=None,
-        max_energy=None,
         mod_update=None,
         seed=None,
         nwalkers=1,
@@ -462,20 +461,20 @@ class WangLandau(MCKernel):
                 An MC step type corresponding to an MCUsher. See valid_mcushers
             bin_size (float):
                 The energy bin size to determine different states.
+            min_energy (float):
+                The minimum energy to sample. Energy value should be given per
+                supercell (i.e. same order of magnitude as what will be sampled).
+            max_energy (float):
+                The maximum energy to sample. Same as above.
             flatness (float): optional
                 The flatness factor used when checking histogram flatness.
                 Must be between 0 and 1.
             mod_factor (float):
                 The modification factor used to update the DOS/entropy.
-                Default is e^1.
+                Default is log(e^1).
             check_period (int): optional
                 The period in number of steps for the histogram flatness to be
                 checked.
-            min_energy (float):
-                The minimum energy to sample. Energy value should be given per
-                supercell (i.e. same order as what will be sampled).
-            max_energy (float):
-                The maximum energy to sample. Same as above.
             mod_update (Callable): optional
                 A function used to update the fill factor when the histogram
                 satisfies the flatness criteria. The function is used to update
@@ -486,9 +485,8 @@ class WangLandau(MCKernel):
             nwalkers (int): optional
                 Number of walkers/chains to sampler. Default is 1.
         """
-        if min_energy is not None and max_energy is not None:
-            if min_energy > max_energy:
-                raise ValueError("min_energy can not be larger than max_energy.")
+        if min_energy > max_energy:
+            raise ValueError("min_energy can not be larger than max_energy.")
         elif mod_factor <= 0:
             raise ValueError("mod_factor must be greater than 0.")
 
@@ -504,19 +502,13 @@ class WangLandau(MCKernel):
         self._bin_size = bin_size
         self._mod_update = mod_update if mod_update is not None else lambda f: f / 2.0
 
-        # default dict of arrays with [entropy, histogram]
-        # keys are generated by _get_key method
-        # nbins = int((max_energy - min_energy) // bin_size)
-        self._aux_states = defaultdict(
-            lambda: np.array(
-                nwalkers
-                * [
-                    [0.0, 0.0],
-                ]
-            )
-        )
-        self._current_energy = np.zeros(nwalkers)
-        self._counter = 0  # count steps to check flatness at given intervals
+        self._energy_levels = np.arange(min_energy, max_energy, bin_size)
+        self._aux_states = {
+            "histogram": np.zeros((nwalkers, len(self._energy_levels)), dtype=int),
[…]
