### 2022-06-09 1dbe2a3
trace cumulative average of features
