+++ smol/moca/sampler/sampler.py
@@ -190,7 +190,7 @@ class Sampler:
         trace = Trace()
         traces = list(map(self._kernel.compute_initial_trace, occupancies))
         for name in traces[0].names:
-            stack = np.vstack([getattr(tr, name) for tr in traces])
+            stack = np.stack([getattr(tr, name) for tr in traces], axis=0)
             setattr(trace, name, stack)
 
         # Initialise progress bar
