### 2022-06-10 50d0019
add histogram update period to WL
