### 2022-06-10 b08c47a
mask zero values in hist and entropy properties
