### 2022-06-10 ce55a4b
mask in energy levels too
