### 2022-06-10 89025c0
floor bin num
