### 2022-06-15 #4356 @MarDiehl
@jarvist: as stated above, I don't feel confident enough to judge the scientific aspect of the submission but would like to follow the review procedure and comment on the technical/programming aspects.
Is there a template for the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) where I could comment on the technical aspects (source code available, test coverage, ...)?
