### 2022-06-24 #4356 @lbluque
Dear @jarvist, I greatly appreciate the time you have already taken to find reviewers.

However, it's been a month since the review could have been started.

I understand JOSS is a volunteer run effort and we all have busy schedules. So if you currently do not have the capacity to serve as editor in a timely manner, please let me know asap so I can see how to proceed.
