### 2022-06-25 #4504 @editorialbot
[REVIEW]: smol: A Python package for cluster expansions and beyond
**Submitting author:** @lbluque (<a href="http://orcid.org/0000-0002-6453-9545">Luis Barroso-Luque</a>)
**Repository:** https://github.com/CederGroupHub/smol
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.1
