### 2022-06-27 f7aa476
Corrected implementation of kernel when bias are present.
