### 2022-06-27 15ecc35
Minor update to test_table_flip_factors in test_mcushers.py.
