### 2022-07-01 CederGroupHub/smol#227 @TomTranter
JOSS review
Hi @lbluque I'm in the process of reviewing smol for your JOSS submission. I will add things here when I find them.
- Installation on windows: After trying to install with "pip install smol" as recommended I received and error for one of the dependencies spglib. I then used conda to install this package from conda-forge instead. The error concerned not having a version of Microsoft Visual C++ even though I do have this. Have you tested the code on Windows? Is this a known issue with spglib?
