### 2022-07-05 CederGroupHub/smol#227 @lbluque
Hi @TomTranter  Thanks a lot for getting this started!

Currently we are not running CI tests for windows. However, I have personally tried to install and use the package on windows without issue.

The `spglib` dependency comes from `pymatgen`. I was able to find this [discussion thread](https://matsci.org/t/installation-problem-building-wheel-for-spglib-pep-517-error/36057), which I think is a similar issue that you had. The suggested workaround is to install `pymatgen` from conda-forge (which is conceptually the same solution you used). But unfortunately I don't think I can address that directly here using `pip`. The "workaround" I can think of is to add a comment describing this in the installation steps.
