### 2022-07-11 aa55f30
Bump pymatgen from 2022.5.26 to 2022.7.8
Bumps [pymatgen](https://github.com/materialsproject/pymatgen) from 2022.5.26 to 2022.7.8.
- [Release notes](https://github.com/materialsproject/pymatgen/releases)
- [Changelog](https://github.com/materialsproject/pymatgen/blob/master/CHANGES.rst)
- [Commits](https://github.com/materialsproject/pymatgen/compare/v2022.5.26...v2022.7.8)

---
updated-dependencies:
- dependency-name: pymatgen
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
