+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy>=1.20.1
 monty==2022.4.26
-pymatgen==2022.5.26
+pymatgen==2022.7.8
