### 2022-07-14 CederGroupHub/smol#227 @TomTranter
Hi @lbluque so actually conda-forge installing just spglib didn't work and you are right I had to install pymatgen, you could introduce a environment.yml for conda installers.
