### 2022-07-14 CederGroupHub/smol#227 @TomTranter
@lbluque the summary in the paper needs to be improved to explain a little more about cluster expansion for a non-specialist
