### 2022-07-17 08ef4df
update install instructions
