+++ docs/src/getting_started.rst
@@ -12,6 +12,13 @@ process headache free. Using ``pip``::
 
         pip install smol
 
+Although **smol** is not tested on Windows platforms, it should still run on Windows
+since there aren't any platform specific dependencies. The only known installation issue
+is building *pymatgen* dependencies. If simply running ``pip install smol`` fails, try
+installing *pymatgen* with conda first::
+
+        conda install -c conda-forge pymatgen
+        pip install smol
 
 Basic Usage
 ===========
