### 2022-07-17 cfc9b10
fix link to notebooks in docs
