+++ binder/requirements.txt
@@ -1,0 +1,5 @@
+numpy>=1.20.1
+monty==2022.4.26
+pymatgen==2022.7.8
+scikit-learn==1.1.1
+h5py==3.7.0
