+++ binder/requirements.txt
@@ -3,3 +3,4 @@ monty==2022.4.26
 pymatgen==2022.7.8
 scikit-learn==1.1.1
 h5py==3.7.0
+crystal_toolkit
