+++ docs/src/getting_started.rst
@@ -155,8 +155,12 @@ package.
 
 Example Notebooks
 =================
+
 For more detailed examples on how to use **smol** have a look at the following
 Jupyter notebooks.
+
+.. image:: https://mybinder.org/badge_logo.svg
+ :target: https://mybinder.org/v2/gh/CederGroupHub/smol/HEAD?labpath=docs%2Fsrc%2Fnotebooks%2Findex.ipynb
 
 Basic Examples
 --------------
