### 2022-07-18 CederGroupHub/smol#227 @lbluque
:wave: @TomTranter thanks so much for taking the time and the detailed suggestions!

I'll mention them individually as I address them.
