### 2022-07-18 CederGroupHub/smol#227 @lbluque
> Hi @lbluque so actually conda-forge installing just spglib didn't work and you are right I had to install pymatgen, you could introduce a environment.yml for conda installers.

Glad to hear installing pymatgen from conda-forge solved the issue. I added additional instructions to the README file and the documentation.

I also added an environment.yml file for user convenience when using conda.
