### 2022-07-18 CederGroupHub/smol#227 @lbluque
> I am now trying to run the examples found here https://cedergrouphub.github.io/smol/notebooks/creating-a-ce.html but the files needed are not to be found in the package so I must clone the repository. To make running the examples easier you could provide a collab workspace and make sure they are all runnable with the files in the repository not just the docs

I created a binder link to be able to run the notebooks online. I added a badge and instructions in the README and documentation.
