+++ paper/paper.md
@@ -41,7 +41,17 @@ bibliography: paper.bib
 
 # Summary
 
-The growing research focus on high principal element materials—which span a variety of applications, such as electrochemical [@Lun:2020], metal alloys [@George:2019], semiconductor, thermoelectric, magnetic, and superconducting [@Gao:2018] materials—necessitates the development of computational methodology capable of resolving atomic configuration behavior and resulting thermodynamic properties. The cluster expansion method coupled with Monte Carlo sampling (CE-MC) is an established and effective way to resolve atomic details underlying important thermodynamic properties [@VanderVen:2018].
+The growing research focus on high principal element materials—which span a
+variety of applications, such as electrochemical [@Lun:2020], metal alloys
+[@George:2019], semiconductor, thermoelectric, magnetic, and superconducting
+[@Gao:2018] materials—necessitates the development of computational methodology
+capable of resolving atomic configuration behavior and resulting thermodynamic
+properties. The cluster expansion method is a formal and effective way to
+construct functions of atomic configuration by coarse-graining materials
+properties, such as formation energies, in terms of species site occupations in
+a lattice model [Sanchez:1984]. The cluster expansion method coupled with Monte
+Carlo sampling (CE-MC) is an established and effective way to resolve atomic
+details underlying important thermodynamic properties [@VanderVen:2018].
 
 `smol` (Statistical Mechanics on Lattices) is a Python package for constructing general applied lattice models, and performing Monte Carlo sampling of associated thermodynamic ensembles. The representation of lattice models in `smol` is based largely on the Cluster Expansion (CE) formalism [@Sanchez:1984], however the package is designed to allow easy implementation of extensions to the formalism, such as redundant representations [@Barroso-Luque:2021]. `smol` also includes flexible and extensible functionality to run Monte Carlo (MC) sampling from canonical and semigrand-canonical ensembles associated with generated lattice models. `smol` has been intentionally designed to be lightweight and include a minimal set of dependencies to enable smooth installation, use and development. `smol` was conceived primarily to enable development and implementation of novel CE-MC methodology but is now sufficiently mature that it is already being used in applied research of real material systems. [@Yang:2022-a; @Chen:2022; @Jadidi:2022; @Yang:2022-b]
 
