### 2022-07-18 CederGroupHub/smol#227 @lbluque
> @lbluque the visualisation example contains no plots. https://cedergrouphub.github.io/smol/notebooks/cluster-visualization.html The Crystal Toolkit Jupyter Lab is not installed on the document rendering website I guess... and I also cannot install it locally to run the example. I'm getting errors on windows with the pip install. It's a shame because I feel the package would really benefit from some visualisation. Perhaps a more basic matplotlib function should be provided as a fall back when this package is not installed. The arrays are hard to interpret by themselves

It is a shame that you were unable to get crystal-toolkit running because it is a great package, and the crystal visualizations look great with it. You should be able to see these and test it out in the Binder I just added.

I also added an example of how to use Plotly (since it is interactive) to visualize a cluster in-lieu of crystal-toolkit. I am hesitant to add this type of functionality directly in our code, since it essentially recreates functionality that already exists within the Materials Project ecosystem that our code is built upon.

I'm afraid I also can't do much in terms of the visualizations showing up in the statically rendering of jupyter right now, since Crystal-Toolkit does not support this yet. See this comment:
https://github.com/CederGroupHub/smol/issues/201#issuecomment-1132142077
