### 2022-07-18 CederGroupHub/smol#227 @lbluque
> @lbluque the summary in the paper needs to be improved to explain a little more about cluster expansion for a non-specialist

I added this explanation to the summary:
"The cluster expansion method is a formal and effective way to construct functions of atomic configuration by coarse-graining materials properties, such as formation energies, in terms of species site occupations in a lattice model [Sanchez:1984]."
