### 2022-07-18 CederGroupHub/smol#227 @lbluque
@TomTranter I think I addressed all your suggestions to some extent now. Please have a look at them when you have some time, and thanks again for taking the time to review our code!
