### 2022-07-20 CederGroupHub/smol#227 @TomTranter
> @lbluque the visualisation example contains no plots. https://cedergrouphub.github.io/smol/notebooks/cluster-visualization.html The Crystal Toolkit Jupyter Lab is not installed on the document rendering website I guess... and I also cannot install it locally to run the example. I'm getting errors on windows with the pip install. It's a shame because I feel the package would really benefit from some visualisation. Perhaps a more basic matplotlib function should be provided as a fall back when this package is not installed. The arrays are hard to interpret by themselves

@lbluque any chance you can fix this visualisation? I think that's it really and everything else looks good to me
