### 2022-07-20 c5585a5
Edited kernel.py to remove single_step from init. Sampler test failed a bit. Trying to fix.
