### 2022-07-20 eee072c
Passed tests. Solves issue #235. Ready to merge.
