### 2022-07-21 923d0b0
update docs to allow offline plotly figures with nbphinx
