+++ getting_started.py
@@ -9,6 +9,16 @@
 # pip install smol
 # ```
 #
+# Although **smol** is not tested on Windows platforms, it should still run on Windows
+# since there aren’t any platform specific dependencies. The only known installation issue
+# is building *pymatgen* dependencies. If simply running `pip install smol` fails, try
+# installing *pymatgen* with conda first:
+#
+# ```
+# conda install -c conda-forge pymatgen
+# pip install smol
+# ```
+#
 # ### Basic Usage
 #
 # **smol** is designed to be simple and intuitive to use. Here is the most
@@ -113,7 +123,9 @@ save_work("CuAu_ce_mc.json", wrangler, expansion, ensemble, sampler.samples)
 # For more detailed examples on how to use **smol** have a look at the following
 # Jupyter notebooks.
 #
-# # Basic Examples
+# You can run the notebooks interactively on Binder.
+#
+# [](https://mybinder.org/v2/gh/CederGroupHub/smol/HEAD?labpath=docs%2Fsrc%2Fnotebooks%2Findex.ipynb)# Basic Examples
 #
 # * Creating a basic cluster expansion
 #
