+++ paper/paper.bib
@@ -282,9 +282,17 @@
 }
 
 @article{Yang:2022-a,
-  title = {Approaches for handling high-dimensional cluster expansions of ionic systems},
-  author = {Yang, Julia H and Chen, Tina and Barroso-Luque, Luis and Jadidi, Zinab and Ceder, Gerbrand},
-  year = {submitted 2022}
+  title = {Approaches for Handling High-Dimensional Cluster Expansions of Ionic Systems},
+  author = {Yang, Julia H. and Chen, Tina and Barroso-Luque, Luis and Jadidi, Zinab and Ceder, Gerbrand},
+  date = {2022-06-27},
+  journaltitle = {npj Computational Materials},
+  shortjournal = {npj Comput Mater},
+  volume = {8},
+  number = {1},
+  pages = {1--11},
+  publisher = {{Nature Publishing Group}},
+  issn = {2057-3960},
+  doi = {10.1038/s41524-022-00818-3}
 }
 
 @article{Yang:2022-b,
