+++ .pre-commit-config.yaml
@@ -38,7 +38,7 @@ repos:
     - --profile=black
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.37.1
+  rev: v2.37.2
   hooks:
     - id: pyupgrade
       args: [--py38-plus]
