### 2022-07-25 6162627
fix valid bin check
