+++ smol/moca/sampler/kernel.py
@@ -642,7 +642,7 @@ class WangLandau(MCKernel):
             self._usher.update_aux_state(step)
 
         # only if bin_num is valid
-        if 0 < bin_num < len(self._energy_levels):
+        if 0 <= bin_num < len(self._energy_levels):
             # compute the cumulative statistics
             total = self._aux_states["total-histogram"][walker, bin_num]
             curr_mean = self._aux_states["mean-features"][walker, bin_num]
