### 2022-07-25 adaaffd
check histogram length in update
