### 2022-07-26 ed4ce15
update citing.rst
