### 2022-08-02 590ae92
add energy predict test to decompprocessor
