+++ smol/cofe/space/orbit.py
@@ -126,20 +126,16 @@ class Orbit(MSONable):
 
     @property
     def multiplicity(self):
-        """Get number of clusters in orbit.
-
-        Number of clusters in the given sites of the lattice object that are in
-        the orbit.
-        """
+        """Get number of clusters in orbit per normilizing unit cell."""
         return len(self.clusters)
 
     @property
     def bit_combos(self):
-        """Get tuple representing the multi-index for site function ordering.
+        """Get tuple of contracted multi-indeces for site function ordering.
 
         tuple of ndarrays, each array is a set of symmetrically equivalent bit
         orderings represented by each row. Bit combos represent non-constant site
-        function orderings, i.e. contracted multi-indices
+        function orderings.
         """
         if self._bit_combos is None:
             # get all the bit symmetry operations
@@ -217,7 +213,7 @@ class Orbit(MSONable):
         """Get the array of correlation functions for all possible configs.
 
         Array of stacked correlation arrays for each symetrically distinct
-        set of bit combos.
+        set of bit combos, i.e a correlation function.
 
         The correlations array is a multidimensional array with each dimension
         corresponding to each site space.
@@ -387,7 +383,7 @@ class Orbit(MSONable):
         return match
 
     def sub_orbit_mappings(self, orbit):
-        """Return a mapping of the sites in the orbit to a sub orbit.
+        """Return a mapping of cluster sites in the orbit to cluster sites in sub orbit.
 
         If the given orbit is not a sub-orbit will return an empty list.
         Note this works for mapping between sites, sites spaces, and basis
@@ -397,7 +393,8 @@ class Orbit(MSONable):
             orbit (Orbit):
                 A sub orbit to return mapping of sites
         Returns:
-            list: of indices sucht that
+            list:
+                a list of indices such that
                 self.base_cluster.sites[indices] = orbit.base_cluster.sites
         """
         indsets = np.array(
@@ -478,7 +475,7 @@ class Orbit(MSONable):
 
         Returns:
             (int, int, int):
-            next orbit id, next bit ordering id, next cluster id
+                next orbit id, next bit ordering id, next cluster id
         """
         self.id = orbit_id
         self.bit_id = orbit_bit_id
