### 2022-08-03 dcc14a6
Changed sampler from single kernel to multiple kernels. Currently using a for loop. Tests revised and passed.
