### 2022-08-03 b0953ad
Adapted to pycodestyle, flake8 and pydocstyle.
