### 2022-08-07 4acf886
Writing new WangLandau kernel based on multi-kernel update. Ongoing.
