+++ smol/moca/sampler/kernel.py
@@ -408,6 +408,219 @@ class Metropolis(ThermalKernel):
         return self.trace
 
 
+class WangLandau(MCKernel):
+    """A Wang-Landau sampling kernel.
+
+    Allows wang-landau sampling method. Stores histogram and DOS in its instance.
+    """
+    valid_mcushers = ALL_MCUSHERS
+    valid_bias = None  # Wang-Landau does not need bias.
+
+    def __init__(
+        self,
+        ensemble,
+        step_type,
+        min_enthalpy,
+        max_enthalpy,
+        bin_size,
+        *args,
+        flatness=0.8,
+        mod_factor=1.0,
+        check_period=1000,
+        update_period=1,
+        mod_update=None,
+        seed=None,
+        **kwargs,
+    ):
+        """Initialize a WangLandau Kernel.
+
+        Args:
+            ensemble (Ensemble):
+                The ensemble object to use to generate samples
+            step_type (str):
+                An MC step type corresponding to an MCUsher. See valid_mcushers
+            min_enthalpy (float):
+                The minimum enthalpy to sample. Enthalpy means the dot product
+                of natural parameters and ensemble features. In canonical ensemble,
+                it shall be the energy. In semigrand ensemble, it shall be E- mu N.
+                Unit is eV scaled at the super-cell.
+            max_enthalpy (float):
+                The maximum enthalpy to sample.
+            bin_size (float):
+                The enthalpy bin size to determine different states.
+            flatness (float): optional
+                The flatness factor used when checking histogram flatness.
+                Must be between 0 and 1.
+            mod_factor (float):
+                log of the initial modification factor used to update the
+                DOS/entropy.
+                Default is 1, which means the DOS shall be multiplied by e^1.
+            check_period (int): optional
+                The period in number of steps for the histogram flatness to be
+                checked.
+            update_period (int): optional
+                The period in number of steps to update the histogram and entropy.
+                In some cases periods slightly larger than each step, may allow
+                more efficient exploration of energy space.
+            mod_update (float|Callable): optional
+                A function used to update the fill factor when the histogram
+                satisfies the flatness criteria. It must monotonically decrease
+                to 0.
+                When a number is provided, at each update we will divide the
+                mod factor by this number.
+            seed (int): optional
+                non-negative integer to seed the PRNG
+            args:
+                Positional arguments to instantiate MCUsher.
+            kwargs:
+                Keyword arguments to instantiate MCUsher.
+        """
+        if min_enthalpy > max_enthalpy:
+            raise ValueError("min_enthalpy can not be larger than max_energy.")
+        if mod_factor <= 0:
+            raise ValueError("mod_factor must be greater than 0.")
+
+        self.flatness = flatness
+        self.check_period = check_period
+        self.update_period = update_period
+        self._m = mod_factor
+        self._window = (min_enthalpy, max_enthalpy, bin_size)
+        if callable(mod_update):
+            self._mod_update = mod_update
+        elif mod_update is not None:
+            self._mod_update = lambda x: x / mod_update
+        else:
+            self._mod_update = lambda x: x / 2.0
+
+        self._levels = np.arange(**self._window)
+
+        # The correct intialization will be handled in set_aux_state.
+        self._current_enthalpy = np.inf
+        self._current_features = np.zeros(len(ensemble.natural_parameters))
+        self._entropy = np.zeros(len(self._levels))
+        self._histogram = np.zeros(len(self._levels), dtype=int)
+        # Used to compute average features only.
+        self._histogram_total = np.zeros(len(self._levels), dtype=int)
[…]
