### 2022-08-07 4974ea4
Tableflip scale factors rewritten to speed up fatorial computations.
