### 2022-08-08 5d5cbd0
Rewritten WangLandau method based on new multi-kernels update. Tests passed.
