+++ smol/moca/sampler/kernel.py
@@ -111,14 +111,14 @@ class MCKernel(ABC):
     valid_bias = None
 
     def __init__(
-        self,
-        ensemble,
-        step_type,
-        *args,
-        seed=None,
-        bias_type=None,
-        bias_kwargs=None,
-        **kwargs,
+            self,
+            ensemble,
+            step_type,
+            *args,
+            seed=None,
+            bias_type=None,
+            bias_kwargs=None,
+            **kwargs,
     ):
         """Initialize MCKernel.
 
@@ -384,10 +384,9 @@ class Metropolis(ThermalKernel):
             self.trace.delta_trace.bias = np.array(
                 self._bias.compute_bias_change(occupancy, step)
             )
-            exponent = (
-                -self.beta * self.trace.delta_trace.enthalpy
-                + self.trace.delta_trace.bias
-            )
+            exponent = (-self.beta * self.trace.delta_trace.enthalpy
+                        + self.trace.delta_trace.bias
+                        )
             self.trace.accepted = np.array(
                 True if exponent >= 0 else exponent > log(self._rng.random())
             )
@@ -396,7 +395,7 @@ class Metropolis(ThermalKernel):
                 True
                 if self.trace.delta_trace.enthalpy <= 0
                 else -self.beta * self.trace.delta_trace.enthalpy
-                > log(self._rng.random())  # noqa
+                     > log(self._rng.random())  # noqa
             )
 
         if self.trace.accepted:
@@ -417,20 +416,20 @@ class WangLandau(MCKernel):
     valid_bias = None  # Wang-Landau does not need bias.
 
     def __init__(
-        self,
-        ensemble,
-        step_type,
-        min_enthalpy,
-        max_enthalpy,
-        bin_size,
-        *args,
-        flatness=0.8,
-        mod_factor=1.0,
-        check_period=1000,
-        update_period=1,
-        mod_update=None,
-        seed=None,
-        **kwargs,
+            self,
+            ensemble,
+            step_type,
+            min_enthalpy,
+            max_enthalpy,
+            bin_size,
+            *args,
+            flatness=0.8,
+            mod_factor=1.0,
+            check_period=1000,
+            update_period=1,
+            mod_update=None,
+            seed=None,
+            **kwargs,
     ):
         """Initialize a WangLandau Kernel.
 
@@ -492,7 +491,7 @@ class WangLandau(MCKernel):
         else:
             self._mod_update = lambda x: x / 2.0
 
-        self._levels = np.arange(**self._window)
+        self._levels = np.arange(min_enthalpy, max_enthalpy, bin_size)
 
         # The correct intialization will be handled in set_aux_state.
         self._current_enthalpy = np.inf
@@ -500,10 +499,13 @@ class WangLandau(MCKernel):
         self._entropy = np.zeros(len(self._levels))
         self._histogram = np.zeros(len(self._levels), dtype=int)
         # Used to compute average features only.
-        self._histogram_total = np.zeros(len(self._levels), dtype=int)
+        self._occurrences = np.zeros(len(self._levels), dtype=int)
         # Mean feature vector per level.
         self._mean_features = np.zeros((len(self._levels),
                                         len(ensemble.natural_parameters)))
+        self._steps_counter = 0  # Log n_steps elapsed.
+        self._states_counter = 0  # Log n_valid_states elapsed
+        # TODO: Do we really need to distinguish them?
 
         # Population of initial trace included here.
         super(WangLandau, self).__init__(ensemble=ensemble,
@@ -513,8 +515,10 @@ class WangLandau(MCKernel):
                                          **kwargs)
         # Additional clean-ups.
         self._histogram[:] = 0
[…]
