### 2022-08-09 b96df57
Minor code style check.
