### 2022-08-10 ba33774
Adjustments on step_counter etc., before merge into CedergroupHub wanglandau branch.
