+++ CHANGES.md
@@ -7,12 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Use this section to keep track of changes in the works.
 ### Added
+### Changed
+### Fixed
+### Removed
+### Deprecated
+
+# [v0.0.5](https://github.com/CederGroupHub/smol/releases/tag/v0.0.5) (2022-08-10)
+### Added
 * Data centering example notebook. #238 @kamronald
 ### Changed
 * Single sampler multiple kernels. #245 @qchempku2017
 ### Fixed
-### Removed
-### Deprecated
+* Fix returning all sub_orbit_mappings. #249 @lbluque
 
 # [v0.0.4](https://github.com/CederGroupHub/smol/releases/tag/v0.0.4) (2022-06-23)
 ### Added
