+++ smol/moca/sampler/sampler.py
@@ -110,7 +110,7 @@ class Sampler:
             )
             for seed in seeds
         ]
-     
+
         # get a trial trace to initialize sample container trace
         _trace = mckernels[0].compute_initial_trace(
             np.zeros(ensemble.num_sites, dtype=int)
