### 2022-08-11 0996db0
remove nwalksers from MCKernel docstring too
