### 2022-08-22 #4504 @jarvist
> > Hi @jarvist, it looks like @kitmccoll has not been very active on github. If you have another means to contact him, can you send him a ping regarding this review please? I worry this has slipped his mind by now.

I've emailed them using the original thread that I used to invite the review. It's a good point that they might not have noticed the prompts if they are not active GitHub users.
