### 2022-08-22 #4504 @lbluque
Thanks a lot @jarvist! I hope we can get an answer relatively soon.

I would find it very helpful to have more transparent communication. It has been quite a frustrating experience to have month-long periods of silence without any statements on timing.
