### 2022-08-29 961da89
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/myint/autoflake: v1.4 → v1.5.1](https://github.com/myint/autoflake/compare/v1.4...v1.5.1)
