### 2022-08-29 a5c86a0
Bump pymatgen from 2022.7.25 to 2022.8.23
Bumps [pymatgen](https://github.com/materialsproject/pymatgen) from 2022.7.25 to 2022.8.23.
- [Release notes](https://github.com/materialsproject/pymatgen/releases)
- [Changelog](https://github.com/materialsproject/pymatgen/blob/master/CHANGES.rst)
- [Commits](https://github.com/materialsproject/pymatgen/compare/v2022.7.25...v2022.8.23)

---
updated-dependencies:
- dependency-name: pymatgen
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
