+++ binder/requirements.txt
@@ -1,7 +1,7 @@
 nbgitpuller
 numpy>=1.20.1
 monty==2022.4.26
-pymatgen==2022.7.25
+pymatgen==2022.8.23
 scikit-learn==1.1.2
 h5py==3.7.0
 crystal_toolkit
