### 2022-09-01 0757dc2
Finished re-implementing comp_space to allow more generic constraints. Testing.
