+++ tests/test_moca/test_comp_space.py
@@ -40,6 +40,49 @@ def comp_space_lmtpo():
                      table_ergodic=True)
 
 
+@pytest.fixture(scope="module")
+def comp_space_lmtpo2():
+    li = Species.from_string('Li+')
+    mn = Species.from_string('Mn3+')
+    ti = Species.from_string('Ti4+')
+    o = Species.from_string('O2-')
+    p = Species.from_string('P3-')
+
+    bits = [[li, mn, ti], [p, o]]
+    sl_sizes = [1, 1]
+    return CompSpace(bits, sl_sizes,
+                     charge_balanced=True,
+                     optimize_basis=True,
+                     table_ergodic=True,
+                     other_constraints=
+                     [([2, 1, 0, 0, 0], 7 / 6)])
+
+
+@pytest.fixture(scope="module")
+def comp_space_lmtpo3():
+    li = Species.from_string('Li+')
+    mn = Species.from_string('Mn3+')
+    ti = Species.from_string('Ti4+')
+    o = Species.from_string('O2-')
+    p = Species.from_string('P3-')
+
+    bits = [[li, mn, ti], [p, o]]
+    sl_sizes = [1, 1]
+    return CompSpace(bits, sl_sizes,
+                     charge_balanced=True,
+                     optimize_basis=True,
+                     table_ergodic=True,
+                     leq_constraints=
+                     [([0, 1, 0, 0, 0], 5 / 6),
+                      ([0, 0, 0, 1, 0], 5 / 6),
+                      ([2, 1, 0, 0, 0], 8 / 6)],
+                     geq_constraints=
+                     [([0, 1, 0, 0, 0], 1 / 6),
+                      ([0, 0, 0, 1, 0], 1 / 6),
+                      ([2, 1, 0, 0, 0], 5 / 6)],
+                     )
+
+
 # A test example with extra constraints.
 @pytest.fixture(scope="module")
 def comp_space_lmntof():
@@ -100,10 +143,14 @@ def test_generic_attributes(comp_space):
     assert np.all(np.any(np.isclose(prim_verts, 0, atol=NUM_TOL), axis=-1))
 
     if comp_space.n_comps_estimate < 10 ** 6:
-        scs = [1, min_sc // 2, min_sc]
+        scs = [1, min_sc // 2, min_sc, min_sc * 2]
         for sc in scs:
-            grid = comp_space.get_comp_grid(sc)
-            ns = grid @ comp_space.basis + comp_space.n0 * sc
+            if not np.allclose(sc * comp_space.b, np.round(sc * comp_space.b)):
+                with pytest.raises(ConstraintIntegerizeError):
+                    grid = comp_space.get_comp_grid(sc)
+            else:
+                grid = comp_space.get_comp_grid(sc)
+            ns = grid @ comp_space.basis + comp_space.get_n0(sc)
             npt.assert_array_equal(grid, comp_space._comp_grids[sc])
             # x-format
             npt.assert_array_equal(A @ ns.T - b[:, None] * sc, 0)
@@ -156,8 +203,13 @@ def test_serialize(comp_space):
                            comp_space._flip_table)
     npt.assert_array_equal(comp_space_reload._vs,
                            comp_space._vs)
-    npt.assert_array_equal(comp_space_reload._n0,
-                           comp_space._n0)
+    for key in comp_space_reload._n0_all:
+        assert key in comp_space._n0_all
+        npt.assert_array_equal(comp_space_reload._n0_all[key],
+                               comp_space._n0_all[key])
+    for key in comp_space._n0_all:
[…]
