+++ smol/moca/comp_space.py
@@ -11,6 +11,7 @@ from pymatgen.core import Element, Composition
 from smol.cofe.space.domain import Vacancy
 
 from .utils.math_utils import (NUM_TOL, get_nonneg_float_vertices,
+                               integerize_vector,
                                integerize_multiple, solve_diophantines,
                                get_natural_solutions, get_optimal_basis,
                                get_ergodic_vectors, get_natural_centroid)
@@ -68,6 +69,21 @@ class ConstraintViolationError(Exception):
         """Represent."""
         return f"Composition: {self.c}, format: {self.form}, " \
                f"tolerance: {NUM_TOL} -> {self.message}"
+
+
+class ConstraintIntegerizeError(Exception):
+    """Raises when constraints cannot be multipled to int under sc_size."""
+
+    def __init__(self, sc_size,
+                 message="Composition constraints cannot be multiplied to integer!"):
+        """Initialize exception class."""
+        self.sc_size = sc_size
+        self.message = message
+        super(ConstraintIntegerizeError, self).__init__(self.message)
+
+    def __str__(self):
+        """Represent."""
+        return f"Super-cell size: {self.sc_size}"
 
 
 class CompUnNormalizedError(Exception):
@@ -176,6 +192,8 @@ class CompSpace(MSONable):
     def __init__(self, bits, sl_sizes=None,
                  charge_balanced=True,
                  other_constraints=None,
+                 leq_constraints=None,
+                 geq_constraints=None,
                  optimize_basis=False,
                  table_ergodic=False):
         """Initialize CompSpace.
@@ -190,10 +208,22 @@ class CompSpace(MSONable):
             charge_balanced(bool): optional
                 Whether to add charge balance constraint. Default
                 to true.
-            other_constraints(List[tuple(1D arrayLike[int], int)]): optional
-                Other integer constraints except charge balance and
-                site-number conservation. Should be given in the form of
-                tuple(a, bb), each gives constraint np.dot(a, n)=bb.
+            other_constraints(List[tuple(1D arrayLike[float], float)]): optional
+                Other equality type composition constraints except charge balance
+                and site-number conservation. Should be given in the form of
+                tuple(a, bb), each gives constraint np.dot(a, n)=bb. a and bb
+                should be in the form of per primitive cell.
+                For example, you may want to constrain n_Li + n_Vac = 0.5 per
+                primitive cell.
+            leq_constraints(List[tuple(1D arrayLike[float], float)]): optional
+                Constraint np.dot(a, n)<=bb. a and bb should be in the form of
+                per primitive cell.
+            geq_constraints(List[tuple(1D arrayLike[float], float)]): optional
+                Constraint np.dot(a, n)>=bb. a and bb should be in the form of
+                per primitive cell.
+                Both leq and geq constraints are only used when enumerating
+                compositions. Table ergodicity code will only consider equality
+                constraints, not leq and geqs.
             optimize_basis(bool): optional
                 Whether to optimize the basis to minimal flip sizes and maximal
                 connectivity in the minimum super-cell size.
@@ -273,20 +303,33 @@ class CompSpace(MSONable):
             if len(a) != self.n_dims:
                 raise ValueError(f"Constraint length: {len(a)} does not match"
                                  f" dimensions: {self.n_dims}!")
-            if not np.allclose(np.array(a), np.round(a)) or not \
-                    np.allclose(b, np.round(b)):
-                raise ValueError("Provided constraints must be integers!")
-            A.append(a)
-            b.append(bb)
[…]

[…]
