### 2022-09-02 92c73cf
Move prim_center out of for loop
Only need to compute once
