+++ smol/moca/__init__.py
@@ -7,6 +7,7 @@ Monte Carlo simulations using Cluster Expansion Hamiltonians.
 
 from smol.moca.sampler.container import SampleContainer
 
+from .comp_space import CompSpace
 from .ensemble.canonical import CanonicalEnsemble
 from .ensemble.ensemble import Ensemble
 from .ensemble.semigrand import SemiGrandEnsemble
@@ -14,7 +15,6 @@ from .processor.composite import CompositeProcessor
 from .processor.ewald import EwaldProcessor
 from .processor.expansion import ClusterExpansionProcessor
 from .sampler.sampler import Sampler
-from .comp_space import CompSpace
 
 __all__ = [
     "ClusterExpansionProcessor",
@@ -25,5 +25,5 @@ __all__ = [
     "SemiGrandEnsemble",
     "Sampler",
     "SampleContainer",
-    "CompSpace"
+    "CompSpace",
 ]
