+++ tests/test_moca/test_comp_space.py
@@ -1,17 +1,24 @@
-import pytest
-import numpy.testing as npt
 import itertools
 import re
 from copy import deepcopy
 
-from smol.moca.comp_space import *
+import numpy.testing as npt
+import pytest
+from pymatgen.core import Composition, Species
+
 from smol.cofe.space.domain import Vacancy
-from pymatgen.core import Species, Composition
-
-from smol.moca.utils.math_utils import (gcd_list, NUM_TOL,
-                                        compute_snf)
+from smol.moca.comp_space import (
+    CompSpace,
+    CompUnNormalizedError,
+    ConstraintIntegerizeError,
+    ConstraintViolationError,
+    NegativeSpeciesError,
+    RoundingError,
+    integerize_vector,
+    np,
+)
+from smol.moca.utils.math_utils import NUM_TOL, compute_snf, gcd_list
 from smol.moca.utils.occu_utils import get_dim_ids_by_sublattice
-
 from tests.utils import assert_msonable, assert_table_set_equal
 
 
@@ -26,82 +33,89 @@ def comp_space(ensemble):
 
 @pytest.fixture(scope="module")
 def comp_space_lmtpo():
-    li = Species.from_string('Li+')
-    mn = Species.from_string('Mn3+')
-    ti = Species.from_string('Ti4+')
-    o = Species.from_string('O2-')
-    p = Species.from_string('P3-')
+    li = Species.from_string("Li+")
+    mn = Species.from_string("Mn3+")
+    ti = Species.from_string("Ti4+")
+    o = Species.from_string("O2-")
+    p = Species.from_string("P3-")
 
     bits = [[li, mn, ti], [p, o]]
     sl_sizes = [1, 1]
-    return CompSpace(bits, sl_sizes,
-                     charge_balanced=True,
-                     optimize_basis=True,
-                     table_ergodic=True)
+    return CompSpace(
+        bits, sl_sizes, charge_balanced=True, optimize_basis=True, table_ergodic=True
+    )
 
 
 @pytest.fixture(scope="module")
 def comp_space_lmtpo2():
-    li = Species.from_string('Li+')
-    mn = Species.from_string('Mn3+')
-    ti = Species.from_string('Ti4+')
-    o = Species.from_string('O2-')
-    p = Species.from_string('P3-')
+    li = Species.from_string("Li+")
+    mn = Species.from_string("Mn3+")
+    ti = Species.from_string("Ti4+")
+    o = Species.from_string("O2-")
+    p = Species.from_string("P3-")
 
     bits = [[li, mn, ti], [p, o]]
     sl_sizes = [1, 1]
-    return CompSpace(bits, sl_sizes,
-                     charge_balanced=True,
-                     optimize_basis=True,
-                     table_ergodic=True,
-                     other_constraints=
-                     [([2, 1, 0, 0, 0], 7 / 6)])
+    return CompSpace(
+        bits,
+        sl_sizes,
+        charge_balanced=True,
+        optimize_basis=True,
+        table_ergodic=True,
+        other_constraints=[([2, 1, 0, 0, 0], 7 / 6)],
+    )
 
 
 @pytest.fixture(scope="module")
 def comp_space_lmtpo3():
-    li = Species.from_string('Li+')
-    mn = Species.from_string('Mn3+')
-    ti = Species.from_string('Ti4+')
-    o = Species.from_string('O2-')
-    p = Species.from_string('P3-')
+    li = Species.from_string("Li+")
+    mn = Species.from_string("Mn3+")
+    ti = Species.from_string("Ti4+")
+    o = Species.from_string("O2-")
+    p = Species.from_string("P3-")
 
     bits = [[li, mn, ti], [p, o]]
     sl_sizes = [1, 1]
-    return CompSpace(bits, sl_sizes,
[…]
