+++ tests/test_moca/test_occu_utils.py
@@ -1,11 +1,17 @@
 """Test smol.moca.utils.occu_utils."""
-import pytest
-
-from smol.moca.utils.occu_utils import *
-import numpy as np
-import numpy.testing as npt
 import itertools
 
+import numpy as np
+import numpy.testing as npt
+import pytest
+
+from smol.moca.utils.occu_utils import (
+    delta_n_from_step,
+    get_dim_ids_by_sublattice,
+    get_dim_ids_table,
+    occu_to_species_list,
+    occu_to_species_n,
+)
 from tests.utils import gen_random_occupancy
 
 
@@ -13,7 +19,9 @@ from tests.utils import gen_random_occupancy
 def sublattices(ensemble):
     sublattices = ensemble.sublattices
     all_sites = [site for sublatt in sublattices for site in sublatt.sites]
-    sites_to_restrict = np.random.choice(all_sites, size=len(all_sites) // 2, replace=False)
+    sites_to_restrict = np.random.choice(
+        all_sites, size=len(all_sites) // 2, replace=False
+    )
     ensemble.restrict_sites(sites_to_restrict)
     return ensemble.sublattices
 
@@ -42,42 +50,43 @@ def test_dim_ids(dim_ids, bits):
     d = sum(len(species) for species in dim_ids)
     for sl_id, dim_id in enumerate(dim_ids):
         assert len(dim_id) == len(bits[sl_id])
-    npt.assert_array_equal(np.arange(d, dtype=int),
-                           list(itertools.chain(*dim_ids)))
+    npt.assert_array_equal(np.arange(d, dtype=int), list(itertools.chain(*dim_ids)))
 
 
 def test_table(sublattices, dim_ids, dim_ids_table_active, dim_ids_table_all):
     # Assert include all sites.
     n_sites = sum(len(sublatt.sites) for sublatt in sublattices)
     n_codes = max(max(sublatt.encoding) for sublatt in sublattices) + 1
-    npt.assert_array_equal(np.sort([site for sublatt in sublattices
-                                    for site in sublatt.sites]),
-                           np.arange(n_sites, dtype=int))
+    npt.assert_array_equal(
+        np.sort([site for sublatt in sublattices for site in sublatt.sites]),
+        np.arange(n_sites, dtype=int),
+    )
     assert dim_ids_table_all.shape == (n_sites, n_codes)
     assert dim_ids_table_active.shape == (n_sites, n_codes)
     for sublatt, dim_ids_sl in zip(sublattices, dim_ids):
-        npt.assert_array_equal(dim_ids_table_all[sublatt.sites[:, None],
-                                                 sublatt.encoding]
-                               - np.array(dim_ids_sl, dtype=int)[None, :], 0)
-        outer_codes = np.setdiff1d(np.arange(n_codes, dtype=int),
-                                   sublatt.encoding).astype(int)
-        npt.assert_array_equal(dim_ids_table_all[sublatt.sites[:, None],
-                                                 outer_codes], -1)
+        npt.assert_array_equal(
+            dim_ids_table_all[sublatt.sites[:, None], sublatt.encoding]
+            - np.array(dim_ids_sl, dtype=int)[None, :],
+            0,
+        )
+        outer_codes = np.setdiff1d(
+            np.arange(n_codes, dtype=int), sublatt.encoding
+        ).astype(int)
+        npt.assert_array_equal(
+            dim_ids_table_all[sublatt.sites[:, None], outer_codes], -1
+        )
         if not sublatt.is_active:
             npt.assert_array_equal(dim_ids_table_active[sublatt.sites, :], -1)
         if sublatt.is_active:
-            npt.assert_array_equal(dim_ids_table_active[sublatt.active_sites
-                                                        [:, None],
-                                                        sublatt.encoding]
[…]
