### 2022-09-02 2ea3f28
set max_lp to max distance from centroid to a site, fix hierarchy unit-tests
