+++ smol/cofe/space/clusterspace.py
@@ -1176,18 +1176,18 @@ class ClusterSubspace(MSONable):
             dict:
                 {size: list of Orbits within diameter cutoff}
         """
-        # Vector sum of a, b, c divided by 2.
         # diameter + max_lp gives maximum possible distance from
         # [0.5, 0.5, 0.5] prim centoid to a point in all enumerable
         # clusters. Add SITE_TOL as a numerical tolerance grace.
         orbits = {1: point_orbits}
-        max_lp = np.linalg.norm(exp_struct.lattice.matrix.sum(axis=0)) / 2
-        max_lp += SITE_TOL
-        prim_center = exp_struct.lattice.get_cartesian_coords([0.5, 0.5, 0.5])
+        centroid = exp_struct.lattice.get_cartesian_coords([0.5, 0.5, 0.5])
+        coords = exp_struct.lattice.get_cartesian_coords(exp_struct.frac_coords)
+        max_lp = max(np.sum((coords - centroid) ** 2, axis=-1) ** 0.5) + SITE_TOL
+
         for size, diameter in sorted(cutoffs.items()):
             new_orbits = []
             neighbors = exp_struct.get_sites_in_sphere(
-                prim_center, diameter + max_lp, include_index=True
+                centroid, diameter + max_lp, include_index=True
             )
             for orbit in orbits[size - 1]:
                 if orbit.base_cluster.diameter > diameter:
