### 2022-09-03 4c6fbcb
use dtype int and bool per numpy deprecation warning
