### 2022-09-03 cacb1c3
add warning filters to tests
