### 2022-09-03 e74c2d8
remove canonical and semigrand ensembles
