+++ smol/moca/ensemble/__init__.py
@@ -1,1 +1,0 @@
-"""Ensemble classes to run MC simulations based on Cluster Expansions."""
