### 2022-09-03 bb11358
update manual build workflow
