+++ smol/moca/__init__.py
@@ -5,6 +5,7 @@ The moca (pronounced mocha) holds implementations of classes used to run
 Monte Carlo simulations using Cluster Expansion Hamiltonians.
 """
 
+from smol.moca.comp_space import CompSpace
 from smol.moca.ensemble import Ensemble
 from smol.moca.processor.composite import CompositeProcessor
 from smol.moca.processor.ewald import EwaldProcessor
