+++ .pre-commit-config.yaml
@@ -18,7 +18,7 @@ repos:
     args: ['--maxkb=500']
 
 - repo: https://github.com/psf/black
-  rev: 22.6.0
+  rev: 22.8.0
   hooks:
   - id: black
 
@@ -44,7 +44,7 @@ repos:
       args: [--py38-plus]
 
 - repo: https://github.com/myint/autoflake
-  rev: v1.5.1
+  rev: v1.5.3
   hooks:
     - id: autoflake
       args:
