### 2022-09-05 393966c
Update spglib requirement from <2.0.0 to <3.0.0
Updates the requirements on [spglib](https://github.com/spglib/spglib) to permit the latest version.
- [Release notes](https://github.com/spglib/spglib/releases)
- [Changelog](https://github.com/spglib/spglib/blob/develop/ChangeLog)
- [Commits](https://github.com/spglib/spglib/compare/v1.8.3.1...v2.0.1)

---
updated-dependencies:
- dependency-name: spglib
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
