### 2022-09-05 #4504 @lbluque
Hi @jarvist checking in again...

Have you heard from @kitmccoll?

If you haven't heard anything, I'd appreciate it if we can find someone else to review.

From the pre-review it seemed MarDiehl was quite interested originally (hopefully that's still the case). https://github.com/openjournals/joss-reviews/issues/4356#issuecomment-1156000093

If not here's the list I had proposed with a few more potential options: https://github.com/openjournals/joss-reviews/issues/4356#issuecomment-1111513178

Thanks!
