### 2022-09-07 f1523cb
More test to resolve issue #261. Improved alias detection algorithm.
