+++ tests/test_cofe/test_clusterspace.py
@@ -282,16 +282,89 @@ def test_orbit_mappings(cluster_subspace, supercell_matrix):
             assert found
 
     # check that the matrix was cached
-    m_hash = tuple(sorted(tuple(s) for s in supercell_matrix))
+    m_hash = tuple(sorted(tuple(s.tolist()) for s in supercell_matrix))
     assert cluster_subspace._supercell_orb_inds[
         m_hash
     ] is cluster_subspace.supercell_orbit_mappings(supercell_matrix)
+
+    # Test that symmetrically equivalent matrices really produce the
+    # same correlation vector for the same occupancy.
+    def get_corr(occu, space, matrix):
+        occu = np.array(occu, dtype=int)
+        mappings = space._gen_orbit_indices(matrix)
+        orbit_list = [
+            (
+                orbit.bit_id,
+                orbit.flat_tensor_indices,
+                orbit.flat_correlation_tensors,
+                cluster_indices,
+            )
+            for cluster_indices, orbit in zip(mappings, space.orbits)
+        ]
+        corr = corr_from_occupancy(occu, space.num_corr_functions, orbit_list)
+        return corr
+
+    structures = [
+        gen_random_structure(cluster_subspace.structure, size=supercell_matrix)
+        for _ in range(10)
+    ]
+    matrix2 = np.array([[0, 1, 0], [0, 0, 1], [1, 0, 0]]) @ np.array(
+        supercell_matrix, dtype=int
+    )
+    matrix3 = np.array([[0, 0, 1], [-1, 0, 0], [0, 1, 0]]) @ np.array(
+        supercell_matrix, dtype=int
+    )
+    # Because of different site_mappings, when you change super-cell,
+    # you must re-generate occupancy strings altogether even if your
+    # super-cell matrices are symmetrically equivalent.
+    occus = [
+        cluster_subspace.occupancy_from_structure(
+            s, scmatrix=supercell_matrix, encode=True
+        )
+        for s in structures
+    ]
+    occus2 = [
+        cluster_subspace.occupancy_from_structure(s, scmatrix=matrix2, encode=True)
+        for s in structures
+    ]
+    occus3 = [
+        cluster_subspace.occupancy_from_structure(s, scmatrix=matrix3, encode=True)
+        for s in structures
+    ]
+    corrs = np.array(
+        [get_corr(occu, cluster_subspace, supercell_matrix) for occu in occus]
+    )
+    corrs2 = np.array([get_corr(occu, cluster_subspace, matrix2) for occu in occus2])
+    corrs3 = np.array([get_corr(occu, cluster_subspace, matrix3) for occu in occus3])
+    # Symmetrically equivalent matrices should give the same correlation
+    # function for the same structure, when the orbit indices mapping
+    # are re-generated.
+    npt.assert_array_almost_equal(corrs, corrs2)
+    npt.assert_array_almost_equal(corrs, corrs3)
+    # Symmetrically equivalent matrices should give the same correlation
+    # vectors on the same structure, when using the default orbit mapping.
+    cluster_subspace._supercell_orb_inds = {}
+    for s, c in zip(structures, corrs):
+        npt.assert_array_almost_equal(
+            cluster_subspace.corr_from_structure(s),
+            cluster_subspace.corr_from_structure(s, scmatrix=supercell_matrix),
+        )
+        npt.assert_array_almost_equal(
+            cluster_subspace.corr_from_structure(s),
+            cluster_subspace.corr_from_structure(s, scmatrix=matrix2),
+        )
+        npt.assert_array_almost_equal(
[…]
