### 2022-09-07 7588eaf
cast enconded occupancy to ndarray of int
