### 2022-09-07 b163f28
method for generating orbit list
