### 2022-09-07 993ab3c
use gen_orbit_list method in procesor (avoid duplicate code)
