### 2022-09-09 #4504 @TomTranter
Hi @arfon, @jarvist, @lbluque I have been following the updates on this post (or lack thereof) and have to say I'm not very impressed by the quality of service being provided by JOSS here to a fellow researcher
