### 2022-09-12 c948387
Bump pymatgen from 2022.8.23 to 2022.9.8
Bumps [pymatgen](https://github.com/materialsproject/pymatgen) from 2022.8.23 to 2022.9.8.
- [Release notes](https://github.com/materialsproject/pymatgen/releases)
- [Changelog](https://github.com/materialsproject/pymatgen/blob/master/CHANGES.rst)
- [Commits](https://github.com/materialsproject/pymatgen/compare/v2022.8.23...v2022.9.8)

---
updated-dependencies:
- dependency-name: pymatgen
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
