### 2022-09-12 c9b31a6
Bump monty from 2022.4.26 to 2022.9.9
Bumps [monty](https://github.com/materialsvirtuallab/monty) from 2022.4.26 to 2022.9.9.
- [Release notes](https://github.com/materialsvirtuallab/monty/releases)
- [Changelog](https://github.com/materialsvirtuallab/monty/blob/master/docs_rst/changelog.rst)
- [Commits](https://github.com/materialsvirtuallab/monty/compare/v2022.4.26...v2022.9.9)

---
updated-dependencies:
- dependency-name: monty
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
