+++ requirements.txt
@@ -1,4 +1,4 @@
 numpy>=1.20.1
-monty==2022.4.26
+monty==2022.9.9
 spglib<2.0.0
 pymatgen==2022.8.23
