+++ binder/requirements.txt
@@ -1,6 +1,6 @@
 nbgitpuller
 numpy>=1.20.1
-monty==2022.4.26
+monty==2022.9.9
 pymatgen==2022.8.23
 scikit-learn==1.1.2
 h5py==3.7.0
