### 2022-09-13 #4504 @jarvist
Thank you Bonan for agreeing to be an additional reviewer on an expedited timeframe of two weeks (27th September).

Bonan, there's some discussion on https://github.com/CederGroupHub/smol/issues/227 from Tom and Luis which may help orientate you to the areas of discussion / focus.
