### 2022-09-14 #4504 @lbluque
> Dear @lbluque and @TomTranter my apologies, I certainly could have been more communicative. I have been attempting to contact Kit weekly, but with no response this side of summer.
>
> I will extend the invitation to review. But also, @kitmccoll, you are very welcome to contribute now if you've been away over summer.

Thank you @jarvist I can certainly understand getting a response from reviewers may be tricky or overly busy schedules can make them fall off the map entirely. As I stated before, it would have just been more helpful to get an update rather than getting no response month-long stretches of time.

In any case, thank you for taking the time to find someone else to review.
