+++ paper/paper.bib
@@ -302,9 +302,9 @@
 }
 
 @article{Chen:2022,
-  title = {Understanding the effects of disorder on the electrochemical properties of LiMn2O4 spinel},
+  title = {Removing the two-phase transition in spinel {LiMn$_2$O$_4$} through cation disorder},
   author = {Chen, Tina and Yang, Julia H and Barroso-Luque, Luis and Ceder, Gerbrand},
-  year = {in prep. 2022}
+  year = {submitted 2022}
 }
 
 @article{Jadidi:2022,
