+++ tests/test_cofe/test_expansion.py
@@ -46,6 +46,16 @@ def test_regression_data(cluster_subspace, rng):
     assert reg_data.parameters == reg.get_params()
     assert_msonable(expansion)
 
+    # test bad feature matrix shape
+    reg_data = RegressionData.from_sklearn(
+        reg, feature_matrix=feat_matrix[:, :-1], property_vector=prop_vec
+    )
+    with pytest.raises(AttributeError):
+        expansion = ClusterExpansion(cluster_subspace, coeffs, reg_data)
+    # test bad coeff length
+    with pytest.raises(AttributeError):
+        expansion = ClusterExpansion(cluster_subspace, coeffs[:-1], reg_data)
+
 
 def test_predict(cluster_expansion, rng):
     subspace = cluster_expansion.cluster_subspace
