### 2022-09-27 f381990
remove _repr utils function
