### 2022-09-27 #4504 @lbluque
> If you are happy to use one of your already published 'Tagged' versions of the repository, it should be quite quick: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

I was not able to figure out how to archive a version already released, so created a new one with a few recent changes. This version still is consistent with the manuscript.
