+++ paper/paper.md
@@ -46,14 +46,14 @@ variety of applications, such as electrochemical [@Lun:2020], structural
 [@George:2019], semiconductor, thermoelectric, magnetic, and superconducting
 [@Gao:2018] materials—necessitates the development of computational methodology
 capable of resolving details of atomic configuration and resulting thermodynamic
-properties. The cluster expansion method is a formal and effective way to
+properties. The cluster expansion (CE) method is a formal and effective way to
 construct functions of atomic configuration by coarse-graining materials
 properties, such as formation energies, in terms of species occupancy
-lattice model [@Sanchez:1984]. The cluster expansion method coupled with Monte
+lattice models [@Sanchez:1984]. The cluster expansion method coupled with Monte
 Carlo sampling (CE-MC) is an established and effective way to resolve atomic
 details underlying important thermodynamic properties [@VanderVen:2018].
 
-`smol` (Statistical Mechanics on Lattices) is a Python package for constructing general applied lattice models, and performing Monte Carlo sampling of associated thermodynamic ensembles. The representation of lattice models in `smol` is based largely on the Cluster Expansion (CE) formalism [@Sanchez:1984]. However, the package is designed to allow easy implementation of extensions to the formalism, such as redundant representations [@Barroso-Luque:2021]. `smol` also includes flexible and extensible functionality to run Monte Carlo (MC) sampling from canonical and semigrand-canonical ensembles associated with the generated lattice models. `smol` has been intentionally designed to be lightweight and include a minimal set of dependencies to enable smooth installation, use, and development. `smol` was conceived primarily to enable development and implementation of novel CE-MC methodology but is now sufficiently mature that it is already being used in applied research of relevant material systems. [@Yang:2022-a; @Chen:2022; @Jadidi:2022; @Yang:2022-b]
+`smol` (Statistical Mechanics on Lattices) is a Python package for constructing generalized applied lattice models, and performing Monte Carlo sampling of associated thermodynamic ensembles. The representation of lattice models in `smol` is based largely on the CE formalism [@Sanchez:1984]. However, the package is designed to allow easy implementation of extensions to the formalism, such as redundant representations [@Barroso-Luque:2021]. `smol` also includes flexible and extensible functionality to run Monte Carlo (MC) sampling from canonical and semigrand-canonical ensembles associated with the generated lattice models. `smol` has been intentionally designed to be lightweight and include a minimal set of dependencies to enable smooth installation, use, and development. `smol` was conceived primarily to enable development and implementation of novel CE-MC methodology but is now sufficiently mature that it is already being used in applied research of relevant material systems. [@Yang:2022-a; @Chen:2022; @Jadidi:2022; @Yang:2022-b]
 
 # Statement of need
 
@@ -63,7 +63,7 @@ Several high-quality software packages implementing CE-MC methodology, such as `
 
 2. `smol` is the only package implemented using `pymatgen`—a widely used Python materials analysis library [@Ong:2013]. This allows seamless use of `pymatgen` functionality for pre and post-processing. Additionally, several other Materials Project [@Jain:2013] packages, such as Fireworks [@Jain:2015], atomate/atomate2 [@Mathew:2017; @atomate2], database creation, and management tools can be leveraged alongside `smol` to include configuration thermodynamic calculations as part of more elaborate materials analysis workflows.
 
-3. `smol` is designed to be intentionally lightweight and dependency lean by delegating much of the non-core functionality to already well-established Python packages, for example, general structure manipulations, enumeration, and linear regression. This makes `smol` easy to install, easy to use, easy to develop, easy to extend and easy to test.
[…]
