### 2022-10-11 #4504 @TomTranter
Congratulations and happy to be part of the process. Nice work on the code

On Thu, 29 Sep 2022 at 18:24, Luis Barroso-Luque ***@***.***>
wrote:

> Thanks @arfon <https://github.com/arfon> ! The proof looks good on my end.
>
